### 2021-10-23 #3847 @whedon
[PRE REVIEW]: Intracranial Electrode Location and Analysis in MNE-Python
**Submitting author:** @alexrockhill (<a href="http://orcid.org/0000-0003-3868-7453">Alexander Rockhill</a>)
**Repository:** <a href="https://github.com/mne-tools/mne-python" target ="_blank">https://github.com/mne-tools/mne-python</a>
**Version:** 0.24.dev0
