### 2021-10-24 ad44473
Add depreciation warning
