+++ mne/preprocessing/nirs/_beer_lambert_law.py
@@ -15,7 +15,7 @@ from ..nirs import source_detector_distances, _channel_frequencies,\
     _check_channels_ordered, _channel_chromophore
 
 
-def beer_lambert_law(raw, ppf=6):
+def beer_lambert_law(raw, ppf=None):
     r"""Convert NIRS optical density data to haemoglobin concentration.
 
     Parameters
@@ -33,6 +33,12 @@ def beer_lambert_law(raw, ppf=6):
     from scipy import linalg
     raw = raw.copy().load_data()
     _validate_type(raw, BaseRaw, 'raw')
+
+    if ppf is None:
+        ppf = 0.1
+        warn('The default value of ppf=0.1 in v0.24 will change to ppf=6 in '
+             'the next MNE-Python release. To utilise the future default '
+             'value set ppf=6.', DeprecationWarning)
 
     freqs = np.unique(_channel_frequencies(raw.info, nominal=True))
     picks = _check_channels_ordered(raw.info, freqs)
