### 2021-10-24 7fafc3c
Replace 'can not' with 'cannot' and improve error messages for dig and custom_ref_applied.
