### 2021-10-24 bf500f3
add back info._check_consistency() [ci skip]
