### 2021-10-24 ec5c8b5
Remove resampling in decoding xdawn example. [circle full]
