+++ examples/decoding/decoding_xdawn_eeg.py
@@ -25,7 +25,7 @@ from sklearn.linear_model import LogisticRegression
 from sklearn.metrics import classification_report, confusion_matrix
 from sklearn.preprocessing import MinMaxScaler
 
-from mne import io, pick_types, read_events, Epochs, EvokedArray
+from mne import io, pick_types, read_events, Epochs, EvokedArray, create_info
 from mne.datasets import sample
 from mne.preprocessing import Xdawn
 from mne.decoding import Vectorizer
@@ -102,11 +102,11 @@ ax.set(ylabel='True label', xlabel='Predicted label')
 fig, axes = plt.subplots(nrows=len(event_id), ncols=n_filter,
                          figsize=(n_filter, len(event_id) * 2))
 fitted_xdawn = clf.steps[0][1]
-tmp = epochs[:1]
-tmp.resample(1)
+info = create_info(epochs.ch_names, 1, epochs.get_channel_types())
+info.set_montage(epochs.get_montage())
 for ii, cur_class in enumerate(sorted(event_id)):
     cur_patterns = fitted_xdawn.patterns_[cur_class]
-    pattern_evoked = EvokedArray(cur_patterns[:n_filter].T, tmp.info, tmin=0)
+    pattern_evoked = EvokedArray(cur_patterns[:n_filter].T, info, tmin=0)
     pattern_evoked.plot_topomap(
         times=np.arange(n_filter),
         time_format='Component %d' if ii == 0 else '', colorbar=False,
