+++ doc/conf.py
@@ -880,8 +880,6 @@ def reset_warnings(gallery_conf, fname):
     warnings.filterwarnings(
         'ignore', message=r'numpy\.ndarray size changed.*',
         category=RuntimeWarning)
-    warnings.filterwarnings(
-        'ignore', message='.*ppf=0.1.*', category=DeprecationWarning)
 
     # In case we use np.set_printoptions in any tutorials, we only
     # want it to affect those:
