### 2021-10-25 85c79d2
Add temp attribute to mne.Info. [ci skip]
