### 2021-10-25 874b34d
add a close function
