### 2021-10-25 1d01f66
TST: Expose failure when resampling to 1 sample with sfreq > len(times) [ci skip]
