### 2021-10-25 218e894
raise a RuntimeError when resampling yields 0 datapoints.
