### 2021-10-25 bfe8490
Change raised error to match numpy error raised in numpy/fft/_pocketfft:94: in _get_backward_norm
