### 2021-10-25 f4deaa8
add basic support for verbose
