### 2021-10-25 f77c172
warn unsupported parameters
