### 2021-10-25 e245405
MAINT: Fix for SciPy pre
