+++ mne/conftest.py
@@ -130,7 +130,6 @@ def pytest_configure(config):
     ignore:.*Passing a schema to Validator.*:DeprecationWarning
     ignore:.*Found the following unknown channel type.*:RuntimeWarning
     ignore:.*in an Any trait will be shared.*:DeprecationWarning
-    ignore:.*ppf=0.1.*:DeprecationWarning
     always::ResourceWarning
     """  # noqa: E501
     for warning_line in warning_lines.split('\n'):
