### 2021-10-25 552723d
return 1 sample instead of raising.
