### 2021-10-25 3a2b7af
reorg docdict to remove incorrect filtering step from create_e*g_epochs funcs
