+++ mne/viz/_brain/_brain.py
@@ -3465,7 +3465,7 @@ class Brain(object):
         The movie is created through the :mod:`imageio` module. The format is
         determined by the extension, and additional options can be specified
         through keyword arguments that depend on the format, see
-        :doc:`imageio's format page <imageio:formats/index>`.
+        :doc:`imageio's format page <imageio:supported_formats>`.
 
         .. Warning::
             This method assumes that time is specified in seconds when adding
