+++ doc/conf.py
@@ -1079,9 +1079,7 @@ def make_redirects(app, exception):
     """Make HTML redirects."""
     # https://www.sphinx-doc.org/en/master/extdev/appapi.html
     # Adapted from sphinxcontrib/redirects (BSD-2-Clause)
-    if not (isinstance(app.builder,
-                       sphinx.builders.html.StandaloneHTMLBuilder) and
-            exception is None):
+    if not isinstance(app.builder, sphinx.builders.html.StandaloneHTMLBuilder):
         return
     logger = sphinx.util.logging.getLogger('mne')
     TEMPLATE = """\
