### 2021-10-25 ec15d64
fix test_repr with different transform.
