### 2021-10-25 44f4b0f
MRG: Change Beer Lambert law default ppf value to 6 (#9843)
* Change Beer Lambert law default ppf value

* Add change log

* Add depreciation warning

* Update mne/preprocessing/nirs/_beer_lambert_law.py

Co-authored-by: Alexandre Gramfort <alexandre.gramfort@m4x.org>

* Explictly set ppf rather than filter warnings

* FIX: Fix tests

* FIX: Dont make two errors

* FIX: Try fix

* Revert "FIX: Try fix"

This reverts commit 984e281801e09cef1be7ec1c017da3fdc47519d9.

* Revert "FIX: Dont make two errors"

This reverts commit 9f2b349e7b1931006231fff1088a8a793929f3f0.

* FIX: Conftest

Co-authored-
[…]
