### 2021-10-25 131c676
MRG, BUG: Make sure resample return at least one sample (#9897)
* TST: Expose failure when resampling to 1 sample with sfreq > len(times) [ci skip]

* raise a RuntimeError when resampling yields 0 datapoints.

* Change raised error to match numpy error raised in numpy/fft/_pocketfft:94: in _get_backward_norm

* fix regex match pattern.

* return 1 sample instead of raising.

* FIX: One more max

* Add entry to latest.inc.

* DOC: Add name

* DOC: Update [ci skip]

Co-authored-by: Eric Larson <larson.eric.d@gmail.com>
