### 2021-10-25 fde89c5
MRG: Add a locking/unlocking context manager to Info for error checking. (#9867)
* Added lock/unlock context manager.

* Update __setitem__ with a dictionary of valid keys and corresponding error msg or checker function.

* Added checkers for bads and description.

* Added try/except/else/finally clause to the contextmanager.

* Replace logger.warning with mne.utils.warn

* fix flake8.

* Added checker functions for user inputs. Only checking type atm.

* Filled a couple of error message examples.

* Add error message including attribute name for every attributes.

* fix 50% of test_info.

* t
[…]
