+++ mne/time_frequency/tfr.py
@@ -640,7 +640,8 @@ def _tfr_aux(method, inst, freqs, decim, return_itc, picks, average,
     out = _compute_tfr(data, freqs, info['sfreq'], method=method,
                        output=output, decim=decim, **tfr_params)
     times = inst.times[decim].copy()
-    info['sfreq'] /= decim.step
+    with info._unlock():
+        info['sfreq'] /= decim.step
 
     if average:
         if return_itc:
