+++ mne/gui/_file_traits.py
@@ -318,6 +318,7 @@ class DigSource(HasPrivateTraits):
             elif len(dir_tree_find(tree, FIFF.FIFFB_ISOTRAK)) > 0:
                 info = _empty_info(1)
                 info['dig'] = read_dig_fif(fname=self.file).dig
+                info._unlocked = False
         else:
             info = read_raw(self.file).info
 
