+++ mne/preprocessing/ica.py
@@ -654,7 +654,8 @@ class ICA(ContainsMixin):
         self.info = pick_info(inst.info, picks)
 
         if self.info['comps']:
-            self.info['comps'] = []
+            with self.info._unlock():
+                self.info['comps'] = []
         self.ch_names = self.info['ch_names']
 
         if isinstance(inst, BaseRaw):
@@ -1075,13 +1076,15 @@ class ICA(ContainsMixin):
         """Aux method."""
         # set channel names and info
         ch_names = []
-        ch_info = info['chs'] = []
+        ch_info = []
         for ii, name in enumerate(self._ica_names):
             ch_names.append(name)
             ch_info.append(dict(
                 ch_name=name, cal=1, logno=ii + 1,
-                coil_type=FIFF.FIFFV_COIL_NONE, kind=FIFF.FIFFV_MISC_CH,
-                coord_frame=FIFF.FIFFV_COORD_UNKNOWN, unit=FIFF.FIFF_UNIT_NONE,
+                coil_type=FIFF.FIFFV_COIL_NONE,
+                kind=FIFF.FIFFV_MISC_CH,
+                coord_frame=FIFF.FIFFV_COORD_UNKNOWN,
+                unit=FIFF.FIFF_UNIT_NONE,
                 loc=np.zeros(12, dtype='f4'),
                 range=1.0, scanno=ii + 1, unit_mul=0))
 
@@ -1091,10 +1094,10 @@ class ICA(ContainsMixin):
             # re-append additionally picked ch_info
             ch_info += [k for k in container.info['chs'] if k['ch_name'] in
                         add_channels]
-        info['bads'] = [ch_names[k] for k in self.exclude]
-        info['projs'] = []  # make sure projections are removed.
-        info._update_redundant()
-        info._check_consistency()
+        with info._unlock(update_redundant=True, check_after=True):
+            info['chs'] = ch_info
+            info['bads'] = [ch_names[k] for k in self.exclude]
+            info['projs'] = []  # make sure projections are removed.
 
     @verbose
     def score_sources(self, inst, target=None, score_func='pearsonr',
