+++ mne/io/ctf/ctf.py
@@ -113,8 +113,9 @@ class RawCTF(BaseRaw):
 
         # Compose a structure which makes fiff writing a piece of cake
         info = _compose_meas_info(res4, coils, coord_trans, eeg)
-        info['dig'] += digs
-        info['dig'] = _format_dig_points(info['dig'])
+        with info._unlock():
+            info['dig'] += digs
+            info['dig'] = _format_dig_points(info['dig'])
         info['bads'] += _read_bad_chans(directory, info)
 
         # Determine how our data is distributed across files
