+++ mne/io/base.py
@@ -1521,7 +1521,7 @@ class BaseRaw(ProjMixin, ContainsMixin, UpdateChannelsMixin, SetChannelsMixin,
              show_first_samp=False, proj=True, group_by='type',
              butterfly=False, decim='auto', noise_cov=None, event_id=None,
              show_scrollbars=True, show_scalebars=True, time_format='float',
-             preload='auto', use_opengl=True, downsampling='auto',
+             precompute='auto', use_opengl=True, downsampling='auto',
              downsampling_method='peak', antialiasing=False,
              overview_mode='channels', verbose=None):
         return plot_raw(self, events, duration, start, n_channels, bgcolor,
@@ -1531,7 +1531,7 @@ class BaseRaw(ProjMixin, ContainsMixin, UpdateChannelsMixin, SetChannelsMixin,
                         proj, group_by, butterfly, decim, noise_cov=noise_cov,
                         event_id=event_id, show_scrollbars=show_scrollbars,
                         show_scalebars=show_scalebars, time_format=time_format,
-                        preload=preload, use_opengl=use_opengl,
+                        precompute=precompute, use_opengl=use_opengl,
                         downsampling=downsampling,
                         downsampling_method=downsampling_method,
                         antialiasing=antialiasing,
