### 2021-10-26 616b943
Supply "title" kwarg to Report in the Report tutorial [skip azp][skip actions] (#9901)
Closes #9848
