+++ tutorials/intro/70_report.py
@@ -82,7 +82,7 @@ subjects_dir = data_path / 'subjects'
 raw_path = sample_dir / 'sample_audvis_filt-0-40_raw.fif'
 raw = mne.io.read_raw(raw_path)
 
-report = mne.Report()
+report = mne.Report(title='Raw example')
 report.add_raw(raw=raw_path, title='Raw from Path')
 report.add_raw(raw=raw, title='Raw from "raw"', psd=False)  # omit PSD plot
 report.save('report_raw.html', overwrite=True)
@@ -98,7 +98,7 @@ events_path = sample_dir / 'sample_audvis_filt-0-40_raw-eve.fif'
 events = mne.read_events(events_path)
 sfreq = raw.info['sfreq']
 
-report = mne.Report()
+report = mne.Report(title='Events example')
 report.add_events(events=events_path, title='Events from Path', sfreq=sfreq)
 report.add_events(events=events, title='Events from "events"', sfreq=sfreq)
 report.save('report_events.html', overwrite=True)
@@ -113,7 +113,7 @@ report.save('report_events.html', overwrite=True)
 
 epochs = mne.Epochs(raw=raw, events=events)
 
-report = mne.Report()
+report = mne.Report(title='Epochs example')
 report.add_epochs(epochs=epochs, title='Epochs from "epochs"')
 report.save('report_epochs.html', overwrite=True)
 
@@ -141,7 +141,7 @@ cov_path = sample_dir / 'sample_audvis-cov.fif'
 evokeds = mne.read_evokeds(evoked_path, baseline=(None, 0))
 evokeds_subset = evokeds[:2]  # The first two
 
-report = mne.Report()
+report = mne.Report(title='Evoked example')
 report.add_evokeds(
     evokeds=evokeds_subset,
     titles=['evoked 1',  # Manually specify titles
@@ -163,7 +163,7 @@ report.save('report_evoked.html', overwrite=True)
 
 cov_path = sample_dir / 'sample_audvis-cov.fif'
 
-report = mne.Report()
+report = mne.Report(title='Covariance example')
 report.add_covariance(cov=cov_path, info=raw_path, title='Covariance')
 report.save('report_cov.html', overwrite=True)
 
@@ -182,7 +182,7 @@ report.save('report_cov.html', overwrite=True)
 ecg_proj_path = sample_dir / 'sample_audvis_ecg-proj.fif'
 eog_proj_path = sample_dir / 'sample_audvis_eog-proj.fif'
 
-report = mne.Report()
+report = mne.Report(title='Projectors example')
 report.add_projs(info=raw_path, title='Projs from info')
 report.add_projs(info=raw_path, projs=ecg_proj_path,
                  title='ECG projs from path')
@@ -250,7 +250,7 @@ eog_components, eog_scores = ica.find_bads_eog(
 )
 ica.exclude = eog_components
 
-report = mne.Report()
+report = mne.Report(title='ICA example')
 report.add_ica(
     ica=ica,
     title='ICA cleaning',
@@ -272,7 +272,7 @@ report.save('report_ica.html', overwrite=True)
 # the resulting file size, you may pass the ``decim`` parameter to only include
 # every n-th volume slice.
 
-report = mne.Report()
+report = mne.Report(title='BEM example')
 report.add_bem(
     subject='sample', subjects_dir=subjects_dir, title='MRI & BEM',
     decim=10
@@ -291,7 +291,7 @@ report.save('report_mri_and_bem.html', overwrite=True)
 
 trans_path = sample_dir / 'sample_audvis_raw-trans.fif'
 
-report = mne.Report()
+report = mne.Report(title='Coregistration example')
 report.add_trans(
     trans=trans_path, info=raw_path, subject='sample',
[…]
