### 2021-10-26 ad82913
Fixes & test coverage
