### 2021-10-26 3b7b28f
TST: show the renderer
