+++ mne/gui/__init__.py
@@ -203,12 +203,11 @@ def coregistration(tabbed=False, split=True, width=None, inst=None,
     if pyvistaqt:
         from ..viz.backends.renderer import MNE_3D_BACKEND_TESTING
         from ._coreg import CoregistrationUI
-        show = not MNE_3D_BACKEND_TESTING
         standalone = not MNE_3D_BACKEND_TESTING
         return CoregistrationUI(
             info_file=inst, subject=subject, subjects_dir=subjects_dir,
             head_resolution=head_high_res, orient_glyphs=orient_to_surface,
-            trans=trans, size=(width, height), show=show, standalone=standalone,
+            trans=trans, size=(width, height), standalone=standalone,
             verbose=verbose
         )
     else:
