### 2021-10-26 118f379
MRG: Small improvements to labels in plot_ica_sources() (#9824)
* Small improvements to labels in plot_ica_sources()

* Fixes & test coverage
