### 2021-10-26 bbf00ce
Revert "TST: show the renderer"
This reverts commit 3b7b28f8c54c7d424a6a4e1eea8586f012595c75.
