### 2021-10-26 1eb4448
TST: use get_config
