### 2021-10-26 0c519d3
Use docdict for more decoding docstrings and add some small fixes & layout improvements [skip azp][skip actions]
