+++ mne/gui/tests/test_coreg_gui.py
@@ -411,17 +411,15 @@ def test_coreg_gui_pyvista(tmpdir, renderer_interactive_pyvistaqt):
     assert coreg._grow_hair == 0
     coreg._set_grow_hair(0.1)
     assert coreg._grow_hair == 0.1
-    orient_to_surface = (config.get('MNE_COREG_ORIENT_TO_SURFACE', '') ==
-                         'true')
-    assert coreg._orient_glyphs == orient_to_surface
+    assert coreg._orient_glyphs == \
+        config.get('MNE_COREG_ORIENT_TO_SURFACE', '') == 'true'
     assert coreg._hpi_coils
     assert coreg._eeg_channels
     assert coreg._head_shape_points
     assert coreg._scale_mode == 'None'
     assert coreg._icp_fid_match == 'nearest'
-    assert not coreg._head_resolution
-    coreg._set_head_resolution(True)
-    assert coreg._head_resolution
+    assert coreg._head_resolution == \
+        config.get('MNE_COREG_HEAD_HIGH_RES', 'true') == 'true'
     assert not coreg._head_transparency
     coreg._set_head_transparency(True)
     assert coreg._head_transparency
