### 2021-10-26 6bcc9a8
restore lighting
