### 2021-10-26 8710097
add _check_fif
