### 2021-10-26 bc3a739
reverse _check_fif logic
