### 2021-10-26 a0d715b
WIP: Deprecate mayavi
