### 2021-10-26 4b47a14
Rpeaks docstring (#9900)
* add docstring note

* improve crossref'd glossary entry

* reorg docdict to remove incorrect filtering step from create_e*g_epochs funcs

* whitespace fix

* better wording [skip github][skip azp]

* flake8

* fix beamformer test

* flake8 again
