### 2021-10-27 fe5ce3b
fix infinite recursion bug
