### 2021-10-27 55442e0
Deprecate ICA.detect_artifacts
Fixes #9906
