### 2021-10-27 b005ea6
MRG: Deprecate ICA.detect_artifacts (#9909)
* Deprecate ICA.detect_artifacts

Fixes #9906

* FIX: Revert extra warning

Co-authored-by: Eric Larson <larson.eric.d@gmail.com>
