+++ environment.yml
@@ -23,6 +23,7 @@ dependencies:
 - spyder-kernels>=1.10.0
 - imageio-ffmpeg>=0.4.1
 - vtk>=9.0.1
+- traitlets
 - pyvista>=0.30
 - pyvistaqt>=0.4
 - qdarkstyle

[…]
