### 2021-10-27 89ed2aa
fix nan channel locations
