### 2021-10-28 ee639af
Small changes to fixed length epochs tutorial (#9259)
* Small changes to fixed length epochs tutorial [skip github]]skip azp]

The tutorial demonstrating `make_fixed_length_epochs()` is titled
"Creating epochs of equal length" in `main`. I always found this
confusing, as Eppchs are *always* supposed to have the same duration.

I've change the title and reworked the first paragraph in an attempt to
make things a little clearer

* adjust title and intro wording

Co-authored-by: Daniel McCloy <dan@mccloy.info>
