+++ tutorials/epochs/60_make_fixed_length_epochs.py
@@ -2,11 +2,15 @@
 """
 .. _tut-fixed-length-epochs:
 
-Creating epochs of equal length
-===============================
+Divide continuous data into equally-spaced epochs
+=================================================
 
-This tutorial shows how to create equal length epochs and briefly demonstrates
-an example of their use in connectivity analysis.
+This tutorial shows how to segment continuous data into a set of epochs spaced
+equidistantly in time. The epochs will not be created based on experimental
+events; instead, the continuous data will be "chunked" into consecutive epochs
+(which may be temporally overlapping, adjacent, or separated).
+We will also briefly demonstrate how to use these epochs in connectivity
+analysis.
 
 First, we import necessary modules and read in a sample raw data set.
 This data set contains brain activity that is event-related, i.e.,
