### 2021-10-28 87f557e
simplify diff
