+++ mne/gui/tests/test_coreg_gui.py
@@ -377,7 +377,7 @@ def test_coreg_gui_pyvista(tmpdir, renderer_interactive_pyvistaqt):
     tempdir = str(tmpdir)
     config = get_config(home_dir=os.environ.get('_MNE_FAKE_HOME_DIR'))
     tmp_trans = op.join(tempdir, 'tmp-trans.fif')
-    coreg = coregistration(inst=None, subject='sample',
+    coreg = coregistration(subject='sample',
                            subjects_dir=subjects_dir, trans=fname_trans)
     assert not coreg._lock_fids
     coreg._reset_fiducials()
