### 2021-10-28 a8de7ca
Add update method using __setitem__.
