### 2021-10-28 b9cded4
FIX: MNE functions using info.update().
