+++ mne/io/hitachi/hitachi.py
@@ -228,8 +228,8 @@ class RawHitachi(BaseRaw):
 
         # Create mne structure
         info = create_info(ch_names, sfreq, ch_types=ch_types)
-        info.update(info_extra)
         with info._unlock():
+            info.update(info_extra)
             info['meas_date'] = meas_date
             for li, loc in enumerate(locs):
                 info['chs'][li]['loc'][:] = loc
