### 2021-10-28 035b68a
FIX: Tests using info.update() [circle full]
