### 2021-10-28 712751e
FIX: Consider argument other as iterable of key/value pairs instead of Mapping. [circle full]
