+++ doc/conf.py
@@ -621,6 +621,8 @@ html_context = {
         '0.12': 'v0.12',
         '0.11': 'v0.11',
     },
+    'switcher_json_url': 'https://mne.tools/dev/_static/versions.json',
+    'switcher_template_url': 'https://mne/tools/{version}/',
     'funders': [
         dict(img='nih.png', size='3', title='National Institutes of Health'),
         dict(img='nsf.png', size='3.5',
