+++ requirements_doc.txt
@@ -1,7 +1,7 @@
 # requirements for building docs
 sphinx!=4.1.0
 https://github.com/numpy/numpydoc/archive/main.zip
-pydata-sphinx-theme>=0.6.3
+https://github.com/drammock/pydata-sphinx-theme/archive/redo-switcher.zip
 https://github.com/sphinx-gallery/sphinx-gallery/archive/master.zip
 sphinxcontrib-bibtex>=2.1.2
 memory_profiler
