### 2021-10-28 cc33859
MAINT: Bump CI image [skip actions] [skip circle]
