### 2021-10-28 dac4f16
Use locking/unlocking for Info.update() (#9914)
* Add update method using __setitem__.

* FIX: MNE functions using info.update().

* FIX: Tests using info.update() [circle full]

* FIX: Consider argument other as iterable of key/value pairs instead of Mapping. [circle full]

* FIX: failing tutorial. [circle full]

* Dont run all [skip azp] [skip actions]

* FIX: tutorial. [skip azp] [skip actions]

Co-authored-by: Eric Larson <larson.eric.d@gmail.com>
