+++ azure-pipelines.yml
@@ -236,9 +236,7 @@ stages:
         ./tools/azure_dependencies.sh
       condition: in(variables['TEST_MODE'], 'pip', 'pip-pre')
       displayName: Install dependencies with pip
-    - script: |
-        call activate base
-        conda env update --name base --file environment.yml
+    - script: conda env update --name base --file environment.yml
       condition: eq(variables['TEST_MODE'], 'conda')
       displayName: Setup MNE environment
     - script: conda remove mne
@@ -299,9 +297,7 @@ stages:
       displayName: Update conda
     - script: conda env update --name base --file environment.yml
       displayName: Setup MNE environment
-    - script: |
-        call activate base
-        conda install -c conda-forge graphviz
+    - script: conda install -c conda-forge graphviz
       displayName: Add graphviz dot
     - script: conda remove --force mne
       displayName: Remove old MNE
