### 2021-10-28 f4b1b8a
Simplify test.
