### 2021-10-28 fd3b643
fix button text on dev version [skip github][skip azp]
