### 2021-10-29 81dfe58
Reorder attributes in docstring in alphabetical order.
