### 2021-10-29 f3500ab
Reorder Notes bullet points in alphabetical order.
