### 2021-10-29 dcab883
Improve couple of error messages in _attributes.
