### 2021-10-29 6d05237
TST: Expose failure with I/O and cropped annotations in Raw.
