### 2021-10-29 cb2ab2d
Add details for fit_params from Notes section to Parameters.
