### 2021-10-29 3a9b066
API: Deprecate mayavi (#9904)
* WIP: Deprecate mayavi

* DOC: Better

* FIX: Dep

* MAINT: Update text

* API: Show warning
