+++ mne/commands/mne_kit2fiff.py
@@ -17,7 +17,9 @@ Use without arguments to invoke GUI:
 
 """
 
+from contextlib import nullcontext
 import sys
+import warnings
 
 import mne
 from mne.io import read_raw_kit
@@ -60,8 +62,15 @@ def run():
 
     input_fname = options.input_fname
     if input_fname is None:
-        with ETSContext():
-            mne.gui.kit2fiff()
+        ctx = nullcontext()
+        try:
+            from mne_kit_gui import kit2fiff  # noqa
+        except ImportError:
+            kit2fiff = mne.gui.kit2fiff
+            ctx = ETSContext()
+        with ctx, warnings.catch_warnings():
+            warnings.simplefilter('always', DeprecationWarning)
+            kit2fiff()
         sys.exit(0)
 
     hsp_fname = options.hsp_fname
