### 2021-10-29 9cfa739
fix coreg instantiation issue with missing input points
