### 2021-10-29 5d1f34c
MRG: Small improvements to ICA documentation. (#9924)
* Add details for fit_params from Notes section to Parameters.

* FIX: Text typo. [skip azp] [skip actions]
