+++ mne/preprocessing/ica.py
@@ -196,7 +196,9 @@ class ICA(ContainsMixin):
         For reference, see :footcite:`Hyvarinen1999,BellSejnowski1995,LeeEtAl1999,AblinEtAl2018`.
     fit_params : dict | None
         Additional parameters passed to the ICA estimator as specified by
-        ``method``.
+        ``method``. Allowed entries are determined by the various algorithm
+        implementations: see :class:`~sklearn.decomposition.FastICA`,
+        :func:`~picard.picard`, :func:`~mne.preprocessing.infomax`.
     max_iter : int | 'auto'
         Maximum number of iterations during fit. If ``'auto'``, it
         will set maximum iterations to ``1000`` for ``'fastica'``
@@ -268,10 +270,9 @@ class ICA(ContainsMixin):
     .. versionchanged:: 0.23
         Warn if `~mne.Epochs` were baseline-corrected.
 
-    .. note:: If you intend to clean fit ICA on `~mne.Epochs`, it is
-              recommended to high-pass filter, but **not** baseline correct the
-              data for good ICA performance. A warning will be emitted
-              otherwise.
+    .. note:: If you intend to fit ICA on `~mne.Epochs`, it is  recommended to
+              high-pass filter, but **not** baseline correct the data for good
+              ICA performance. A warning will be emitted otherwise.
 
     A trailing ``_`` in an attribute name signifies that the attribute was
     added to the object during fitting, consistent with standard scikit-learn
