+++ mne/tests/test_cov.py
@@ -193,9 +193,9 @@ def _assert_reorder(cov_new, cov_orig, order):
                     cov_orig['data'], atol=1e-20)
 
 
-def test_ad_hoc_cov(tmpdir):
+def test_ad_hoc_cov(tmp_path):
     """Test ad hoc cov creation and I/O."""
-    out_fname = tmpdir.join('test-cov.fif')
+    out_fname = tmp_path.join('test-cov.fif')
     evoked = read_evokeds(ave_fname)[0]
     cov = make_ad_hoc_cov(evoked.info)
     cov.save(out_fname)
@@ -218,13 +218,13 @@ def test_ad_hoc_cov(tmpdir):
         cov._get_square()
 
 
-def test_io_cov(tmpdir):
+def test_io_cov(tmp_path):
     """Test IO for noise covariance matrices."""
     cov = read_cov(cov_fname)
     cov['method'] = 'empirical'
     cov['loglik'] = -np.inf
-    cov.save(tmpdir.join('test-cov.fif'))
-    cov2 = read_cov(tmpdir.join('test-cov.fif'))
+    cov.save(tmp_path.join('test-cov.fif'))
+    cov2 = read_cov(tmp_path.join('test-cov.fif'))
     assert_array_almost_equal(cov.data, cov2.data)
     assert_equal(cov['method'], cov2['method'])
     assert_equal(cov['loglik'], cov2['loglik'])
@@ -232,24 +232,24 @@ def test_io_cov(tmpdir):
 
     cov2 = read_cov(cov_gz_fname)
     assert_array_almost_equal(cov.data, cov2.data)
-    cov2.save(tmpdir.join('test-cov.fif.gz'))
-    cov2 = read_cov(tmpdir.join('test-cov.fif.gz'))
+    cov2.save(tmp_path.join('test-cov.fif.gz'))
+    cov2 = read_cov(tmp_path.join('test-cov.fif.gz'))
     assert_array_almost_equal(cov.data, cov2.data)
 
     cov['bads'] = ['EEG 039']
     cov_sel = pick_channels_cov(cov, exclude=cov['bads'])
     assert cov_sel['dim'] == (len(cov['data']) - len(cov['bads']))
     assert cov_sel['data'].shape == (cov_sel['dim'], cov_sel['dim'])
-    cov_sel.save(tmpdir.join('test-cov.fif'))
+    cov_sel.save(tmp_path.join('test-cov.fif'))
 
     cov2 = read_cov(cov_gz_fname)
     assert_array_almost_equal(cov.data, cov2.data)
-    cov2.save(tmpdir.join('test-cov.fif.gz'))
-    cov2 = read_cov(tmpdir.join('test-cov.fif.gz'))
+    cov2.save(tmp_path.join('test-cov.fif.gz'))
+    cov2 = read_cov(tmp_path.join('test-cov.fif.gz'))
     assert_array_almost_equal(cov.data, cov2.data)
 
     # test warnings on bad filenames
-    cov_badname = tmpdir.join('test-bad-name.fif.gz')
+    cov_badname = tmp_path.join('test-bad-name.fif.gz')
     with pytest.warns(RuntimeWarning, match='-cov.fif'):
         write_cov(cov_badname, cov)
     with pytest.warns(RuntimeWarning, match='-cov.fif'):
@@ -257,7 +257,7 @@ def test_io_cov(tmpdir):
 
 
 @pytest.mark.parametrize('method', (None, 'empirical', 'shrunk'))
-def test_cov_estimation_on_raw(method, tmpdir):
+def test_cov_estimation_on_raw(method, tmp_path):
     """Test estimation from raw (typically empty room)."""
     if method == 'shrunk':
         try:
@@ -301,8 +301,8 @@ def test_cov_estimation_on_raw(method, tmpdir):
[…]
