+++ mne/tests/test_morph.py
@@ -224,7 +224,7 @@ def assert_power_preserved(orig, new, limits=(1., 1.05)):
 
 @requires_h5py
 @testing.requires_testing_data
-def test_surface_vector_source_morph(tmpdir):
+def test_surface_vector_source_morph(tmp_path):
     """Test surface and vector source estimate morph."""
     inverse_operator_surf = read_inverse_operator(fname_inv_surf)
 
@@ -256,9 +256,9 @@ def test_surface_vector_source_morph(tmpdir):
     assert 'surface' in repr(source_morph_surf)
 
     # check loading and saving for surf
-    source_morph_surf.save(tmpdir.join('42.h5'))
+    source_morph_surf.save(tmp_path.join('42.h5'))
 
-    source_morph_surf_r = read_source_morph(tmpdir.join('42.h5'))
+    source_morph_surf_r = read_source_morph(tmp_path.join('42.h5'))
 
     assert (all([read == saved for read, saved in
                  zip(sorted(source_morph_surf_r.__dict__),
@@ -279,7 +279,7 @@ def test_surface_vector_source_morph(tmpdir):
 @requires_dipy()
 @pytest.mark.slowtest
 @testing.requires_testing_data
-def test_volume_source_morph_basic(tmpdir):
+def test_volume_source_morph_basic(tmp_path):
     """Test volume source estimate morph, special cases and exceptions."""
     import nibabel as nib
     inverse_operator_vol = read_inverse_operator(fname_inv_vol)
@@ -342,14 +342,14 @@ def test_volume_source_morph_basic(tmpdir):
                              subjects_dir=subjects_dir)
 
     # two different ways of saving
-    source_morph_vol.save(tmpdir.join('vol'))
+    source_morph_vol.save(tmp_path.join('vol'))
 
     # check loading
-    source_morph_vol_r = read_source_morph(tmpdir.join('vol-morph.h5'))
+    source_morph_vol_r = read_source_morph(tmp_path.join('vol-morph.h5'))
 
     # check for invalid file name handling ()
     with pytest.raises(IOError, match='not found'):
-        read_source_morph(tmpdir.join('42'))
+        read_source_morph(tmp_path.join('42'))
 
     # check morph
     stc_vol_morphed = source_morph_vol.apply(stc_vol)
@@ -473,7 +473,7 @@ def test_volume_source_morph_basic(tmpdir):
         ('sample', 'fsaverage', 7.0, 7.4, float, True),  # morph_mat
     ])
 def test_volume_source_morph_round_trip(
-        tmpdir, subject_from, subject_to, lower, upper, dtype, morph_mat,
+        tmp_path, subject_from, subject_to, lower, upper, dtype, morph_mat,
         monkeypatch):
     """Test volume source estimate morph round-trips well."""
     import nibabel as nib
@@ -563,7 +563,7 @@ def test_volume_source_morph_round_trip(
     stc_from_unit_rt = morph_to_from.apply(morph_from_to.apply(stc_from_unit))
     assert_power_preserved(stc_from_unit, stc_from_unit_rt, limits=limits)
     if morph_mat:
-        fname = tmpdir.join('temp-morph.h5')
+        fname = tmp_path.join('temp-morph.h5')
         morph_to_from.save(fname)
         morph_to_from = read_source_morph(fname)
         assert morph_to_from.vol_morph_mat is None
@@ -745,7 +745,7 @@ def test_morph_stc_sparse():
     (slice(1, 2), 51, 204, 12),
     (slice(0, 2), 88, 324, 20),
 ])
-def test_volume_labels_morph(tmpdir, sl, n_real, n_mri, n_orig):
[…]
