+++ mne/datasets/sleep_physionet/tests/test_physionet.py
@@ -24,7 +24,7 @@ pooch = _soft_import('pooch', 'dataset downloading', strict=True)
 
 
 @pytest.fixture(scope='session')
-def physionet_tmp_path(tmp_path_factory):
+def physionet_tmpdir(tmp_path_factory):
     """Fixture exposing a temporary directory for testing."""
     return str(tmp_path_factory.mktemp('physionet_files'))
 
@@ -88,63 +88,63 @@ def test_run_update_age_records(tmp_path):
 
 
 @pytest.mark.parametrize('subject', [39, 68, 69, 78, 79, 83])
-def test_sleep_physionet_age_missing_subjects(physionet_tmp_path, subject,
+def test_sleep_physionet_age_missing_subjects(physionet_tmpdir, subject,
                                               download_is_error):
     """Test handling of missing subjects in Sleep Physionet age fetcher."""
     with pytest.raises(
             ValueError, match='This dataset contains subjects 0 to 82'):
         age.fetch_data(
             subjects=[subject], recording=[1], on_missing='raise',
-            path=physionet_tmp_path)
+            path=physionet_tmpdir)
     with pytest.warns(RuntimeWarning,
                       match='This dataset contains subjects 0 to 82'):
         age.fetch_data(
             subjects=[subject], recording=[1], on_missing='warn',
-            path=physionet_tmp_path)
+            path=physionet_tmpdir)
     paths = age.fetch_data(
         subjects=[subject], recording=[1], on_missing='ignore',
-        path=physionet_tmp_path)
+        path=physionet_tmpdir)
     assert paths == []
 
 
 @pytest.mark.parametrize('subject,recording', [(13, 2), (36, 1), (52, 1)])
-def test_sleep_physionet_age_missing_recordings(physionet_tmp_path, subject,
+def test_sleep_physionet_age_missing_recordings(physionet_tmpdir, subject,
                                                 recording, download_is_error):
     """Test handling of missing recordings in Sleep Physionet age fetcher."""
     with pytest.raises(
             ValueError, match=f'Requested recording {recording} for subject'):
         age.fetch_data(subjects=[subject], recording=[recording],
-                       on_missing='raise', path=physionet_tmp_path)
+                       on_missing='raise', path=physionet_tmpdir)
     with pytest.warns(RuntimeWarning,
                       match=f'Requested recording {recording} for subject'):
         age.fetch_data(subjects=[subject], recording=[recording],
-                       on_missing='warn', path=physionet_tmp_path)
+                       on_missing='warn', path=physionet_tmpdir)
     paths = age.fetch_data(subjects=[subject], recording=[recording],
-                           on_missing='ignore', path=physionet_tmp_path)
+                           on_missing='ignore', path=physionet_tmpdir)
     assert paths == []
 
 
-def test_sleep_physionet_age(physionet_tmp_path, monkeypatch, download_is_error):
+def test_sleep_physionet_age(physionet_tmpdir, monkeypatch, download_is_error):
     """Test Sleep Physionet URL handling."""
     # check download_is_error patching
     with pytest.raises(AssertionError, match='Test should not download'):
-        age.fetch_data(subjects=[0], recording=[1], path=physionet_tmp_path)
+        age.fetch_data(subjects=[0], recording=[1], path=physionet_tmpdir)
     # then patch
     my_func = _FakeFetch()
[…]
