+++ azure-pipelines.yml
@@ -242,7 +242,7 @@ stages:
     - script: conda env update --name base --file environment.yml
       condition: eq(variables['TEST_MODE'], 'conda')
       displayName: Setup MNE environment
-    - script: conda remove mne
+    - script: conda remove -yq --force mne
       condition: eq(variables['TEST_MODE'], 'conda')
       displayName: Remove old MNE
     - script: pip install -e .
@@ -302,7 +302,7 @@ stages:
       displayName: Setup MNE environment
     - script: conda install -c conda-forge graphviz
       displayName: Add graphviz dot
-    - script: conda remove --force mne
+    - script: conda remove -yq --force mne
       displayName: Remove old MNE
     - script: pip install -e .
       displayName: Install dev MNE
