### 2021-10-29 47c0112
Explain blender shrinkwrap operation (#9921)
* Explain blender shrinkwrap operation

* reflow text to fix too-long lines

* Update tutorials/forward/80_fix_bem_in_blender.py

* Update tutorials/forward/80_fix_bem_in_blender.py

* Update tutorials/forward/80_fix_bem_in_blender.py

Co-authored-by: Daniel McCloy <dan@mccloy.info>
Co-authored-by: Eric Larson <larson.eric.d@gmail.com>
