### 2021-10-29 bbd6641
Fix some joins
