### 2021-10-29 ccb6ada
Fix test_annotations.py
