+++ mne/io/fiff/tests/test_raw_fiff.py
@@ -83,7 +83,7 @@ def test_acq_skip(tmp_path):
     assert_allclose(data, np.concatenate(expected_data, axis=-1), atol=1e-22)
 
     # Check that acquisition skips are handled properly during I/O
-    fname = tmp_path.join('test_raw.fif')
+    fname = tmp_path / 'test_raw.fif'
     raw.save(fname, fmt=raw.orig_format)
     # first: file size should not increase much (orig data is missing
     # 7 of 17 buffers, so if we write them out it should increase the file
@@ -181,7 +181,7 @@ def test_subject_info(tmp_path):
     for key, val in zip(keys, vals):
         subject_info[key] = val
     raw.info['subject_info'] = subject_info
-    out_fname = tmp_path.join('test_subj_info_raw.fif')
+    out_fname = tmp_path / 'test_subj_info_raw.fif'
     raw.save(out_fname, overwrite=True)
     raw_read = read_raw_fif(out_fname)
     for key in keys:
@@ -213,7 +213,7 @@ def test_output_formats(tmp_path):
     # let's fake a raw file with different formats
     raw = read_raw_fif(test_fif_fname).crop(0, 1)
 
-    temp_file = tmp_path.join('raw.fif')
+    temp_file = tmp_path / 'raw.fif'
     for ii, (fmt, tol) in enumerate(zip(formats, tols)):
         # Let's test the overwriting error throwing while we're at it
         if ii > 0:
@@ -253,7 +253,7 @@ def test_multiple_files(tmp_path):
     # going in reverse order so the last fname is the first file (need later)
     raws = [None] * len(tmins)
     for ri in range(len(tmins) - 1, -1, -1):
-        fname = tmp_path.join('test_raw_split-%d_raw.fif' % ri)
+        fname = tmp_path / ('test_raw_split-%d_raw.fif' % ri)
         raw.save(fname, tmin=tmins[ri], tmax=tmaxs[ri])
         raws[ri] = read_raw_fif(fname)
         assert (len(raws[ri].times) ==
@@ -393,11 +393,11 @@ def test_split_files(tmp_path, mod, monkeypatch):
     # Test a very close corner case
 
     assert_allclose(raw_1.buffer_size_sec, 10., atol=1e-2)  # samp rate
-    split_fname = tmp_path.join(f'split_raw_{mod}.fif')
+    split_fname = tmp_path / f'split_raw_{mod}.fif'
     # intended filenames
-    split_fname_elekta_part2 = tmp_path.join(f'split_raw_{mod}-1.fif')
-    split_fname_bids_part1 = tmp_path.join(f'split_raw_split-01_{mod}.fif')
-    split_fname_bids_part2 = tmp_path.join(f'split_raw_split-02_{mod}.fif')
+    split_fname_elekta_part2 = tmp_path / f'split_raw_{mod}-1.fif'
+    split_fname_bids_part1 = tmp_path / f'split_raw_split-01_{mod}.fif'
+    split_fname_bids_part2 = tmp_path / f'split_raw_split-02_{mod}.fif'
     raw_1.set_annotations(Annotations([2.], [5.5], 'test'))
 
     # Check that if BIDS is used and no split is needed it defaults to
@@ -515,15 +515,15 @@ def test_split_files(tmp_path, mod, monkeypatch):
     assert op.isdir(tmp_path)
     with pytest.raises(ValueError, match='must end with an underscore'):
         raw_crop.save(
-            tmp_path.join('test.fif'), split_naming='bids', verbose='error')
+            tmp_path / 'test.fif', split_naming='bids', verbose='error')
 
[…]
