### 2021-10-30 92c1da1
adjust topo layout to better fill space; ensure legend axes do not overlap data axes
