### 2021-10-30 421f304
Fix small coreg bugs for CTF compat
