+++ mne/gui/_coreg.py
@@ -6,9 +6,12 @@ import os.path as op
 import numpy as np
 from traitlets import observe, HasTraits, Unicode, Bool, Float
 
+from ..io.constants import FIFF
 from ..defaults import DEFAULTS
-from ..io import read_info, read_fiducials
+from ..io import read_info, read_fiducials, read_raw
 from ..io.pick import pick_types
+from ..io.open import fiff_open, dir_tree_find
+from ..io.meas_info import _empty_info
 from ..coreg import Coregistration, _is_mri_subject
 from ..viz._3d import (_plot_head_surface, _plot_head_fiducials,
                        _plot_head_shape_points, _plot_mri_fiducials,
@@ -16,6 +19,7 @@ from ..viz._3d import (_plot_head_surface, _plot_head_fiducials,
 from ..transforms import (read_trans, write_trans, _ensure_trans,
                           rotation_angles, _get_transforms_to_coord_frame)
 from ..utils import get_subjects_dir, check_fname, _check_fname, fill_doc, warn
+from ..channels import read_dig_fif
 
 
 @fill_doc
@@ -142,7 +146,7 @@ class CoregistrationUI(HasTraits):
         )
 
         # process requirements
-        info = read_info(info_file) if info_file is not None else None
+        info = None
         subjects_dir = get_subjects_dir(
             subjects_dir=subjects_dir, raise_error=True)
         subject = _get_default(subject, self._get_subjects(subjects_dir)[0])
@@ -333,7 +337,19 @@ class CoregistrationUI(HasTraits):
 
     @observe("_info_file")
     def _info_file_changed(self, change=None):
-        self._info = read_info(self._info_file)
+        if not self._info_file:
+            return
+        elif self._info_file.endswith(('.fif', '.fif.gz')):
+            fid, tree, _ = fiff_open(self._info_file)
+            fid.close()
+            if len(dir_tree_find(tree, FIFF.FIFFB_MEAS_INFO)) > 0:
+                self._info = read_info(self._info_file, verbose=False)
+            elif len(dir_tree_find(tree, FIFF.FIFFB_ISOTRAK)) > 0:
+                self._info = _empty_info(1)
+                self._info['dig'] = read_dig_fif(fname=self._info_file).dig
+                self._info._unlocked = False
+        else:
+            self._info = read_raw(self._info_file).info
         # XXX: add coreg.set_info()
         self._coreg._info = self._info
         self._coreg._setup_digs()
@@ -602,12 +618,16 @@ class CoregistrationUI(HasTraits):
         if self._eeg_channels:
             eeg = ["original"]
             picks = pick_types(self._info, eeg=(len(eeg) > 0))
-            eeg_actors = _plot_sensors(
-                self._renderer, self._info, self._to_cf_t, picks, meg=False,
-                eeg=eeg, fnirs=False, warn_meg=False, head_surf=self._head_geo,
-                units='m', sensor_opacity=self._defaults["sensor_opacity"],
-                orient_glyphs=self._orient_glyphs, surf=self._head_geo)
-            eeg_actors = eeg_actors["eeg"]
+            if len(picks) > 0:
+                eeg_actors = _plot_sensors(
+                    self._renderer, self._info, self._to_cf_t, picks,
+                    meg=False, eeg=eeg, fnirs=False, warn_meg=False,
+                    head_surf=self._head_geo, units='m',
+                    sensor_opacity=self._defaults["sensor_opacity"],
+                    orient_glyphs=self._orient_glyphs, surf=self._head_geo)
+                eeg_actors = eeg_actors["eeg"]
+            else:
+                eeg_actors = None
         else:
             eeg_actors = None
         self._update_actor("eeg_channels", eeg_actors)
