+++ mne/tests/test_source_space.py
@@ -172,7 +172,7 @@ def test_add_source_space_distances_limited(tmp_path):
     n_do = 200  # limit this for speed
     src_new[0]['vertno'] = src_new[0]['vertno'][:n_do].copy()
     src_new[1]['vertno'] = src_new[1]['vertno'][:n_do].copy()
-    out_name = tmp_path.join('temp-src.fif')
+    out_name = tmp_path / 'temp-src.fif'
     add_source_space_distances(src_new, dist_limit=0.007)
     write_source_spaces(out_name, src_new)
     src_new = read_source_spaces(out_name)
@@ -206,7 +206,7 @@ def test_add_source_space_distances(tmp_path):
     n_do = 19  # limit this for speed
     src_new[0]['vertno'] = src_new[0]['vertno'][:n_do].copy()
     src_new[1]['vertno'] = src_new[1]['vertno'][:n_do].copy()
-    out_name = tmp_path.join('temp-src.fif')
+    out_name = tmp_path / 'temp-src.fif'
     n_jobs = 2
     assert n_do % n_jobs != 0
     with pytest.raises(ValueError, match='non-negative'):
@@ -247,9 +247,9 @@ def test_discrete_source_space(tmp_path):
     v = src[0]['vertno']
 
     # let's make a discrete version with the C code, and with ours
-    temp_name = tmp_path.join('temp-src.fif')
+    temp_name = tmp_path / 'temp-src.fif'
     # save
-    temp_pos = tmp_path.join('temp-pos.txt')
+    temp_pos = tmp_path / 'temp-pos.txt'
     np.savetxt(str(temp_pos), np.c_[src[0]['rr'][v], src[0]['nn'][v]])
     # let's try the spherical one (no bem or surf supplied)
     run_subprocess(['mne_volume_source_space', '--meters',
@@ -288,7 +288,7 @@ def test_discrete_source_space(tmp_path):
 def test_volume_source_space(tmp_path):
     """Test setting up volume source spaces."""
     src = read_source_spaces(fname_vol)
-    temp_name = tmp_path.join('temp-src.fif')
+    temp_name = tmp_path / 'temp-src.fif'
     surf = read_bem_surfaces(fname_bem, s_id=FIFF.FIFFV_BEM_SURF_ID_BRAIN)
     surf['rr'] *= 1e3  # convert to mm
     bem_sol = read_bem_solution(fname_bem_3_sol)
@@ -343,7 +343,7 @@ def test_other_volume_source_spaces(tmp_path):
     # Travis doesn't seem to like them
 
     # let's try the spherical one (no bem or surf supplied)
-    temp_name = tmp_path.join('temp-src.fif')
+    temp_name = tmp_path / 'temp-src.fif'
     run_subprocess(['mne_volume_source_space',
                     '--grid', '7.0',
                     '--src', temp_name,
@@ -463,7 +463,7 @@ def test_setup_source_space(tmp_path):
 
     # oct-6 (sample) - auto filename + IO
     src = read_source_spaces(fname)
-    temp_name = tmp_path.join('temp-src.fif')
+    temp_name = tmp_path / 'temp-src.fif'
     with pytest.warns(None):  # sklearn equiv neighbors
         src_new = setup_source_space('sample', spacing='oct6',
                                      subjects_dir=subjects_dir, add_dist=False)
@@ -491,7 +491,7 @@ def test_setup_source_space(tmp_path):
 @pytest.mark.parametrize('spacing', [2, 7])
 def test_setup_source_space_spacing(tmp_path, spacing):
     """Test setting up surface source spaces using a given spacing."""
-    copytree(op.join(subjects_dir, 'sample'), str(tmp_path.join('sample')))
[…]
