### 2021-10-30 1badf4b
Read haemoglobin data from SNIRF
