+++ mne/io/snirf/_snirf.py
@@ -98,14 +98,17 @@ class RawSNIRF(BaseRaw):
                      "MNE does not support this feature. "
                      "Only the first dataset will be processed.")
 
-            if np.array(dat.get('nirs/data1/measurementList1/dataType')) != 1:
-                raise RuntimeError('File does not contain continuous wave '
-                                   'data. MNE only supports reading continuous'
-                                   ' wave amplitude SNIRF files. Expected type'
-                                   ' code 1 but received type code %d' %
-                                   (np.array(dat.get(
-                                       'nirs/data1/measurementList1/dataType'
-                                   ))))
+            snirf_data_type = np.array(dat.get('nirs/data1/measurementList1'
+                                               '/dataType'))
+            if np.array(dat.get('nirs/data1/measurementList1/dataType')) \
+                    not in [1, 99999]:
+                # 1 = Continous Wave
+                # 99999 = Processed
+                raise RuntimeError('MNE only supports reading continuous'
+                                   ' wave amplitude and processed haemoglobin'
+                                   ' SNIRF files. Expected type'
+                                   ' code 1 or 99999 but received type '
+                                   f'code {snirf_data_type}')
 
             last_samps = dat.get('/nirs/data1/dataTimeSeries').shape[0] - 1
 
@@ -151,24 +154,29 @@ class RawSNIRF(BaseRaw):
 
             # Source and detector labels are optional fields.
             # Use S1, S2, S3, etc if not specified.
-            if 'sourceLabels' in dat['nirs/probe']:
+            if 'sourceLabels_disabled' in dat['nirs/probe']:
+                # This is disabled as
+                # MNE-Python does not currently support custom source names.
+                # Instead, sources must be integer values.
                 sources = np.array(dat.get('nirs/probe/sourceLabels'))
                 sources = [s.decode('UTF-8') for s in sources]
             else:
-                sources = np.unique([dat.get('nirs/data1/' + c +
-                                             '/sourceIndex')[0]
-                                     for c in channels])
-                sources = [f"S{s}" for s in sources]
+                sources = np.unique([_correct_shape(np.array(dat.get(
+                    'nirs/data1/' + c + '/sourceIndex')))[0]
+                    for c in channels])
+                sources = [f"S{int(s)}" for s in sources]
 
-            if 'detectorLabels' in dat['nirs/probe']:
+            if 'detectorLabels_disabled' in dat['nirs/probe']:
+                # This is disabled as
+                # MNE-Python does not currently support custom detector names.
+                # Instead, detector must be integer values.
                 detectors = np.array(dat.get('nirs/probe/detectorLabels'))
                 detectors = [d.decode('UTF-8') for d in detectors]
-
             else:
-                detectors = np.unique([dat.get('nirs/data1/' + c +
-                                               '/detectorIndex')[0]
-                                       for c in channels])
-                detectors = [f"D{d}" for d in detectors]
+                detectors = np.unique([_correct_shape(np.array(dat.get(
+                    'nirs/data1/' + c + '/detectorIndex')))[0]
+                    for c in channels])
+                detectors = [f"D{int(d)}" for d in detectors]
 
             # Extract source and detector locations
             # 3D positions are optional in SNIRF,
@@ -207,26 +215,42 @@ class RawSNIRF(BaseRaw):
             assert len(detectors) == detPos3D.shape[0]
 
             chnames = []
[…]
