### 2021-10-30 d9988eb
Read haemoglobin data from SNIRF
