### 2021-10-30 3ebf0ad
Read haemoglobin data from SNIRF
