+++ mne/tests/test_source_estimate.py
@@ -438,7 +438,7 @@ def test_io_stc_h5(tmp_path, is_complex, vector):
     else:
         stc = _fake_stc(is_complex=is_complex)
     pytest.raises(ValueError, stc.save, tmp_path / 'tmp', ftype='foo')
-    out_name = tmp_path / 'tmp'
+    out_name = str(tmp_path / 'tmp')
     stc.save(out_name, ftype='h5')
     stc.save(out_name, ftype='h5')  # test overwrite
     stc3 = read_source_estimate(out_name)
