### 2021-10-30 343ba1a
Extend GUI testing
