+++ tutorials/io/30_reading_fnirs_data.py
@@ -49,6 +49,25 @@ using :func:`mne.io.read_raw_snirf`.
           wave or haemoglobin data stored in the .snirf format.
 
 
+Specifying the coordinate system
+--------------------------------
+
+There is a variety of coordinate systems used to specify the location of
+sensors (see :ref:`tut-source-alignment` for details). Where possible the
+coordinate system will be determined automatically when reading a SNIRF file.
+However, sometimes this is not possible and you must manually specify the
+coordinate frame the optodes are in. This is done using the ``optode_frame``
+argument when loading data.
+
+=======  ==================  =================
+Vendor   Coordinate Frame    ``optode_frame``
+=======  ==================  =================
+NIRx     ICBM-152 MNI        mri
+Kernel   Freesurfer MRI      mri
+=======  ==================  =================
+
+
+
 ***********************
 Continuous Wave Devices
 ***********************
