### 2021-10-31 f7715ce
rearange GUI test - saves a few seconds.
