### 2021-10-31 5d911d7
remove automatic installation of mne-qt-browser
