### 2021-10-31 92fccdc
revert removal of automatic installation until mne-qt-browser is uploaded to PyPi
