### 2021-11-01 9cbff29
Coreg fixes (#9928)
* fix coreg instantiation issue with missing input points

* Fix small coreg bugs for CTF compat

* Extend GUI testing

* rearange GUI test - saves a few seconds.
