+++ mne/gui/_coreg.py
@@ -6,9 +6,13 @@ import os.path as op
 import numpy as np
 from traitlets import observe, HasTraits, Unicode, Bool, Float
 
+from ..io.constants import FIFF
 from ..defaults import DEFAULTS
-from ..io import read_info, read_fiducials
+from ..io import read_info, read_fiducials, read_raw
 from ..io.pick import pick_types
+from ..io.open import fiff_open, dir_tree_find
+from ..io.meas_info import _empty_info
+from ..io._read_raw import supported as raw_supported_types
 from ..coreg import Coregistration, _is_mri_subject
 from ..viz._3d import (_plot_head_surface, _plot_head_fiducials,
                        _plot_head_shape_points, _plot_mri_fiducials,
@@ -16,6 +20,7 @@ from ..viz._3d import (_plot_head_surface, _plot_head_fiducials,
 from ..transforms import (read_trans, write_trans, _ensure_trans,
                           rotation_angles, _get_transforms_to_coord_frame)
 from ..utils import get_subjects_dir, check_fname, _check_fname, fill_doc, warn
+from ..channels import read_dig_fif
 
 
 @fill_doc
@@ -142,7 +147,7 @@ class CoregistrationUI(HasTraits):
         )
 
         # process requirements
-        info = read_info(info_file) if info_file is not None else None
+        info = None
         subjects_dir = get_subjects_dir(
             subjects_dir=subjects_dir, raise_error=True)
         subject = _get_default(subject, self._get_subjects(subjects_dir)[0])
@@ -215,10 +220,23 @@ class CoregistrationUI(HasTraits):
     def _set_info_file(self, fname):
         if fname is None:
             return
-        if not self._check_fif('info', fname):
+
+        # info file can be anything supported by read_raw
+        try:
+            check_fname(fname, 'info', tuple(raw_supported_types.keys()),
+                        endings_err=tuple(raw_supported_types.keys()))
+        except IOError as e:
+            warn(e)
+            self._widgets["info_file"].set_value(0, '')
             return
-        self._info_file = _check_fname(
-            fname, overwrite=True, must_exist=True, need_dir=False)
+
+        # ctf ds `files` are actually directories
+        if fname.endswith(('.ds',)):
+            self._info_file = _check_fname(
+                fname, overwrite=True, must_exist=True, need_dir=True)
+        else:
+            self._info_file = _check_fname(
+                fname, overwrite=True, must_exist=True, need_dir=False)
 
     def _set_omit_hsp_distance(self, distance):
         self._omit_hsp_distance = distance
@@ -333,7 +351,19 @@ class CoregistrationUI(HasTraits):
 
     @observe("_info_file")
     def _info_file_changed(self, change=None):
-        self._info = read_info(self._info_file)
+        if not self._info_file:
+            return
+        elif self._info_file.endswith(('.fif', '.fif.gz')):
+            fid, tree, _ = fiff_open(self._info_file)
+            fid.close()
+            if len(dir_tree_find(tree, FIFF.FIFFB_MEAS_INFO)) > 0:
+                self._info = read_info(self._info_file, verbose=False)
+            elif len(dir_tree_find(tree, FIFF.FIFFB_ISOTRAK)) > 0:
+                self._info = _empty_info(1)
+                self._info['dig'] = read_dig_fif(fname=self._info_file).dig
+                self._info._unlocked = False
+        else:
+            self._info = read_raw(self._info_file).info
         # XXX: add coreg.set_info()
         self._coreg._info = self._info
         self._coreg._setup_digs()
@@ -602,12 +632,16 @@ class CoregistrationUI(HasTraits):
         if self._eeg_channels:
             eeg = ["original"]
[…]
