### 2021-11-01 a5e930a
BUG: Fix bug with Mayavi backend
