### 2021-11-01 0873f13
Improve random_state docstring & some other small doc fixes [skip azp][skip actions]
