### 2021-11-01 0654593
add mne-qt-browser to requirements.txt
