### 2021-11-01 8d51563
FIX: Index and plot
