### 2021-11-01 0daf2ed
Attempt to make "verbose" docstring more use-friendly [skip azp][skip actions]
