### 2021-11-01 f8283e8
MRG: Improve random_state docstring & some other small doc fixes (#9934)
* Improve random_state docstring & some other small doc fixes [skip azp][skip actions]

* Add missing space [skip actions][skip azp]

* Forgot to commit a line [skip azp][skip actions]
