### 2021-11-01 9ff25f3
New attempt to slightly improve verbose docstring [skip azp][skip actions]
