### 2021-11-01 93da058
New attempt to slightly improve verbose docstring [skip azp][skip actions] (#9936)
