### 2021-11-01 e8dd72b
MRG, MAINT: Bump CI image (#9917)
* MAINT: Bump CI image [skip actions] [skip circle]

* Try again [skip actions] [skip circle]

* FIX: Another [skip actions] [skip circle]

* FIX: Another [skip actions] [skip circle]

* FIX: Another [skip actions] [skip circle]

* FIX: Better [skip actions] [skip circle]

* FIX: Again [skip actions] [skip circle]

* FIX: One more [skip actions] [skip circle]

* FIX: One more [skip actions] [skip circle]

* FIX: One more [skip actions] [skip circle]

* FIX: One more [skip actions] [skip circle]

* Dont build all [skip actions] [skip circle]

* Dont build all [
[…]
