### 2021-11-01 68c1b0c
Use _check_fname() instead of str()
