+++ mne/tests/test_rank.py
@@ -266,7 +266,7 @@ def test_maxfilter_get_rank(n_proj, fname, rank_orig, meg, tol_kind, tol):
 
 def test_explicit_bads_pick():
     """Test when bads channels are explicitly passed + default picks=None."""
-    raw = read_raw_fif(raw_fname, preload=True)
+    raw = read_raw_fif(raw_fname).crop(0, 5).load_data()
     raw.pick_types(eeg=True, meg=True, ref_meg=True)
 
     # Covariance
