### 2021-11-01 bf945f1
[BUG, MRG] Fix infinite recursion bug in `get_montage_volume_labels` (#9907)
* fix infinite recursion bug

* add PR number

* update test

* fix function reference
