+++ mne/surface.py
@@ -1923,7 +1923,7 @@ _VOXELS_MAX = 100  # define constant to avoid runtime issues
 
 @fill_doc
 def get_montage_volume_labels(montage, subject, subjects_dir=None,
-                              aseg='aparc+aseg', dist=5):
+                              aseg='aparc+aseg', dist=2):
     """Get regions of interest near channels from a Freesurfer parcellation.
 
     .. note:: This is applicable for channels inside the brain
@@ -2081,5 +2081,5 @@ def _voxel_neighbors(seed, image, thresh=None, max_peak_dist=1,
             if len(voxels) > voxels_max:
                 break
             next_neighbors = next_neighbors.union(voxel_neighbors)
-        neighbors = next_neighbors
+        neighbors = next_neighbors.difference(voxels)
     return voxels
