### 2021-11-01 784a769
Fix more joins
