+++ mne/minimum_norm/tests/test_inverse.py
@@ -378,7 +378,7 @@ def test_localization_bias_loose(bias_params_fixed, method, lower, upper,
         ('sLORETA', 93, 95, 0.95, 0.96, {}, 0.8, 1),
         pytest.param('eLORETA', 93, 100, 0.95, 0.96,
                      dict(method_params=dict(force_equal=True)), None, 1,
-                     marks = pytest.mark.slowtest),
+                     marks=pytest.mark.slowtest),
         pytest.param('eLORETA', 100, 100, 0.98, 0.99, {}, None, 1.0,
                      marks=pytest.mark.slowtest),
         pytest.param('eLORETA', 100, 100, 0.98, 0.99, {}, 0.8, 1.0,
