+++ tools/github_actions_dependencies.sh
@@ -28,6 +28,8 @@ else
 	pip install --progress-bar off https://github.com/pyvista/pyvistaqt/zipball/main
 	echo "imageio-ffmpeg, xlrd, mffpy"
 	pip install --progress-bar off --pre imageio-ffmpeg xlrd mffpy
+	echo "mne-qt-browser"
+	pip install --progress_bar off --pre mne-qt-browser
 fi
 pip install --progress-bar off --upgrade -r requirements_testing.txt
 if [ "${DEPS}" != "minimal" ]; then
