### 2021-11-01 f97a3ce
fix typo in github_actions_dependencies.sh
