### 2021-11-01 0b513d4
fix docstring for plot_raw again
