### 2021-11-01 f7c4f1f
fix docstring for plot_raw
