### 2021-11-02 f8746f1
Replace more tmpdir with tmp_path
