### 2021-11-02 9337e18
add mne-qt-browser to azure_dependencies.sh
