+++ mne/viz/raw.py
@@ -209,8 +209,8 @@ def plot_raw(raw, events=None, duration=10.0, start=0.0, n_channels=20,
     By default, the channel means are removed when ``remove_dc`` is set to
     ``True``. This flag can be toggled by pressing 'd'.
 
-    (IPython + pyqtgraph) You need to set the magic command
-     ``%%gui qt5`` in IPython for the pyqtgraph-backend to work.
+    .. note:: For the pyqtgraph-backend to run in IPython with ``block=False``
+              you need to enter the magic command ``%%gui qt5`` once before.
     """
     from ..io.base import BaseRaw
     from ._figure import _get_browser
