### 2021-11-02 1c80e6a
add mne-qt-browser to environment.yml
