+++ environment.yml
@@ -42,3 +42,4 @@ dependencies:
 - pooch
 - pip:
   - ipyvtklink
+  - mne-qt-browser
