### 2021-11-02 7b4da67
Update mne/viz/_figure.py
Co-authored-by: Eric Larson <larson.eric.d@gmail.com>
