### 2021-11-02 b8b49ab
CI: Speed up Azure tests (#9937)
* WIP: Figure out why epochs slow on win [skip actions] [skip circle]

* FIX: Modernize [skip actions] [skip circle]

* FIX: More [skip actions] [skip circle]

* FIX: More [skip circle] [skip actions]

* FIX: More [skip actions] [skip circle]

* FIX: Last one
