+++ .circleci/config.yml
@@ -26,8 +26,9 @@ jobs:
             name: Complete checkout
             command: |
               if ! git remote -v | grep upstream; then
-                git remote add upstream git://github.com/mne-tools/mne-python.git
+                git remote add upstream https://github.com/mne-tools/mne-python.git
               fi
+              git remote set-url upstream https://github.com/mne-tools/mne-python.git
               git fetch upstream
         - save_cache:
             key: source-cache
