### 2021-11-02 a1d4eb9
MRG, BUG: Fix bugs with snapshot_brain_montage (#9933)
* BUG: Fix bug with Mayavi backend

* FIX: Index and plot
