### 2021-11-02 3eca0ad
FIX: fiducials locking in coreg gui (#9913)
* fix

* simplify diff

* simplify diff

* update testing
