+++ tools/azure_dependencies.sh
@@ -16,7 +16,6 @@ elif [ "${TEST_MODE}" == "pip-pre" ]; then
 	python -m pip install --progress-bar off --upgrade --pre --only-binary "vtk" vtk
 	python -m pip install --progress-bar off https://github.com/pyvista/pyvista/zipball/main
 	python -m pip install --progress-bar off https://github.com/pyvista/pyvistaqt/zipball/main
-	python -m pip install --progress-bar off --upgrade --pre mne-qt-browser
 else
 	echo "Unknown run type ${TEST_MODE}"
 	exit 1
