+++ mne/viz/_figure.py
@@ -657,8 +657,7 @@ def set_browser_backend(backend_name, verbose=None):
        +--------------------------------------+------------+-----------+
        | :func:`plot_ica_sources`             | ✓          |           |
        +--------------------------------------+------------+-----------+
-       +--------------------------------------+------------+-----------+
-       | **Feature:**                                                  |
+       | **Feature:**                         |            |           |
        +--------------------------------------+------------+-----------+
        | Show Events                          | ✓          | ✓         |
        +--------------------------------------+------------+-----------+
