+++ mne/viz/_figure.py
@@ -657,7 +657,8 @@ def set_browser_backend(backend_name, verbose=None):
        +--------------------------------------+------------+-----------+
        | :func:`plot_ica_sources`             | ✓          |           |
        +--------------------------------------+------------+-----------+
-       | **Feature:**                         |            |           |
+       +--------------------------------------+------------+-----------+
+       | **Feature:**                                                  |
        +--------------------------------------+------------+-----------+
        | Show Events                          | ✓          | ✓         |
        +--------------------------------------+------------+-----------+
@@ -674,7 +675,7 @@ def set_browser_backend(backend_name, verbose=None):
        | Overview-Bar (with Z-Score-Mode)     |            | ✓         |
        +--------------------------------------+------------+-----------+
 
-       .. versionadded:: 0.24
+    .. versionadded:: 0.24
     """
     global MNE_BROWSE_BACKEND
     old_backend_name = MNE_BROWSE_BACKEND
