### 2021-11-02 6962002
DOC: Add link to mne-qt-browser issues in docs
