### 2021-11-02 a5f3db7
BUG: Force mayavi in test that uses it
