### 2021-11-02 34a4966
Update mne/utils/docs.py
Co-authored-by: Daniel McCloy <dan@mccloy.info>
