### 2021-11-02 f350a7f
Update mne/utils/docs.py
Co-authored-by: Daniel McCloy <dan@mccloy.info>
