### 2021-11-02 368876d
Update mne/viz/raw.py
Co-authored-by: Daniel McCloy <dan@mccloy.info>
