+++ mne/viz/backends/_qt.py
@@ -653,6 +653,10 @@ class _Renderer(_PyVistaRenderer, _QtDock, _QtToolBar, _QtMenuBar,
                 self.plotter.app_window.show()
         self._update()
         for plotter in self._all_plotters:
+            # After being shown, the sizes can be wrong for example if a
+            # dock is added. Let's send the actual size to each plotter
+            sz = plotter.size()
+            plotter.window_size = (sz.width(), sz.height())
             plotter._render()
         self._process_events()
 
