### 2021-11-02 1221b9c
BUG: Force mayavi in test that uses it (#9944)
* BUG: Force mayavi in test that uses it

* BUG: Fix appearance and bug

* FIX: Geom
