### 2021-11-02 b97da43
Replace tmpdir with tmp_path (#9925)
* Replace tmpdir with tmp_path

* Fix flake8

* Fix some joins

* Fix test_annotations.py

* More path join fixes

* Even more fixes

* More fixes

* Hopefully final set of fixes

* More fun with tmp_dir.joins

* Missing )

* More...

* Moar!

* Fix test

* Use _check_fname() instead of str()

* Forgot argument

* Set overwrite=True

* Fix more joins

* Missing parens

* Replace more tmpdir with tmp_path

* Convert Path to str

* Overwrite

* Forgot merge conflict

* FIX: Specific

Co-authored-by: Eric Larson <larson.eric.d@gmail.com>
