+++ mne/io/tests/test_write.py
@@ -10,9 +10,9 @@ from mne.io.constants import FIFF
 from mne.io.write import start_file, write_int
 
 
-def test_write_int(tmpdir):
+def test_write_int(tmp_path):
     """Test that write_int raises an error on bad values."""
-    with start_file(tmpdir.join('temp.fif')) as fid:
+    with start_file(tmp_path / 'temp.fif') as fid:
         write_int(fid, FIFF.FIFF_MNE_EVENT_LIST, [2147483647])  # 2 ** 31 - 1
         write_int(fid, FIFF.FIFF_MNE_EVENT_LIST, [])  # 2 ** 31 - 1
         with pytest.raises(TypeError, match=r'.*exceeds max.*EVENT_LIST\)'):
