+++ mne/datasets/sleep_physionet/tests/test_physionet.py
@@ -24,9 +24,9 @@ pooch = _soft_import('pooch', 'dataset downloading', strict=True)
 
 
 @pytest.fixture(scope='session')
-def physionet_tmpdir(tmpdir_factory):
+def physionet_tmpdir(tmp_path_factory):
     """Fixture exposing a temporary directory for testing."""
-    return str(tmpdir_factory.mktemp('physionet_files'))
+    return str(tmp_path_factory.mktemp('physionet_files'))
 
 
 class _FakeFetch:
@@ -78,10 +78,10 @@ def _check_mocked_function_calls(mocked_func, call_fname_hash_pairs,
 @requires_good_network
 @requires_pandas
 @requires_version('xlrd', '0.9')
-def test_run_update_age_records(tmpdir):
+def test_run_update_age_records(tmp_path):
     """Test Sleep Physionet URL handling."""
     import pandas as pd
-    fname = op.join(str(tmpdir), "records.csv")
+    fname = op.join(str(tmp_path), "records.csv")
     _update_sleep_age_records(fname)
     data = pd.read_csv(fname)
     pd.testing.assert_frame_equal(data, pd.read_csv(AGE_SLEEP_RECORDS))
@@ -178,10 +178,10 @@ def test_sleep_physionet_age(physionet_tmpdir, monkeypatch, download_is_error):
 @requires_good_network
 @requires_pandas
 @requires_version('xlrd', '0.9')
-def test_run_update_temazepam_records(tmpdir):
+def test_run_update_temazepam_records(tmp_path):
     """Test Sleep Physionet URL handling."""
     import pandas as pd
-    fname = op.join(str(tmpdir), "records.csv")
+    fname = op.join(str(tmp_path), "records.csv")
     _update_sleep_temazepam_records(fname)
     data = pd.read_csv(fname)
 
