### 2021-11-03 5bafe52
[ENH]: New pyqtgraph-backend for 2D-Data-Browser (#9687)
* start abstraction of Browser-Classes

* outline of data-management-class

* wip refactoring MNEBrowserFigure

* refactored Browser-Initialization into MNEDataBrowser

* fix some style-issues and remove remnants

* start with integration of pyqtgraph-prototype

* make pyqtgraph optional (just for development, can be removed again for final PR)

* more refactoring of mpl-methods into BrowserBase

* fixes for failing tests

* add docstrings in _browser.py

* move _update_projector-call into BrowserBase

* add use/set/get for browser

* re
[…]
