+++ tools/github_actions_infos.sh
@@ -1,4 +1,5 @@
 #!/bin/bash -ef
 
+which mne
 mne sys_info -pd
 python -c "import numpy; numpy.show_config()"
