+++ doc/visualization.rst
@@ -79,3 +79,6 @@ Visualization
    close_3d_figure
    close_all_3d_figures
    get_brain_class
+   set_browser_backend
+   get_browser_backend
+   use_browser_backend
