+++ requirements.txt
@@ -37,3 +37,4 @@ mffpy>=0.5.7
 ipywidgets
 ipyvtklink
 pooch
+mne-qt-browser

[…]
