+++ doc/conf.py
@@ -617,8 +617,9 @@ xxl = '6'
 html_context = {
     'build_dev_html': bool(int(os.environ.get('BUILD_DEV_HTML', False))),
     'versions_dropdown': {
-        'dev': 'v0.24 (devel)',
-        'stable': 'v0.23 (stable)',
+        'dev': 'v0.25 (devel)',
+        'stable': 'v0.24 (stable)',
+        '0.23': 'v0.23',
         '0.22': 'v0.22',
         '0.21': 'v0.21',
         '0.20': 'v0.20',
