+++ doc/whats_new.rst
@@ -5,7 +5,7 @@ What's new
 
 .. currentmodule:: mne
 
-.. include:: changes/latest.inc
+.. include:: changes/0.24.inc
 .. include:: changes/0.23.inc
 .. include:: changes/0.22.inc
 .. include:: changes/0.21.inc
