+++ mne/_version.py
@@ -3,4 +3,4 @@
 #
 # License: BSD-3-Clause
 
-__version__ = '0.24.0'
+__version__ = '1.0.dev0'
