### 2021-11-03 6b54015
MAINT: Remove Mayavi
