+++ mne/report/tests/test_report.py
@@ -350,7 +350,7 @@ def test_report_raw_psd_and_date(tmp_path):
     assert 'Unknown' in ''.join(report.html)
 
 
-@pytest.mark.slowtest  # slow for Mayavi on Azure
+@pytest.mark.slowtest  # slow on Azure
 @testing.requires_testing_data
 def test_render_add_sections(renderer, tmp_path):
     """Test adding figures/images to section."""
@@ -778,7 +778,7 @@ def test_manual_report_2d(tmp_path, invisible_fig):
     r.save(fname=fname, open_browser=False)
 
 
-@pytest.mark.slowtest  # 30 sec on Azure for Mayavi
+@pytest.mark.slowtest  # 30 sec on Azure
 @testing.requires_testing_data
 def test_manual_report_3d(tmp_path, renderer):
     """Simulate adding 3D sections."""

[…]
