+++ mne/utils/__init__.py
@@ -42,7 +42,7 @@ from ._testing import (run_command_if_main, requires_sklearn,
                        requires_version, requires_nibabel, requires_mne,
                        requires_good_network, requires_pandas, requires_h5py,
                        ArgvSetter, SilenceStdout, has_freesurfer, has_mne_c,
-                       _TempDir, has_nibabel, _import_mlab, buggy_mkl_svd,
+                       _TempDir, has_nibabel, buggy_mkl_svd,
                        requires_numpydoc, requires_vtk, requires_freesurfer,
                        requires_nitime, requires_dipy,
                        requires_neuromag2ft, requires_pylsl,
