+++ mne/utils/check.py
@@ -659,8 +659,11 @@ def _check_option(parameter, value, allowed_values, extra=''):
         options = f'The only allowed value is {repr(allowed_values[0])}'
     else:
         options = 'Allowed values are '
-        options += ', '.join([f'{repr(v)}' for v in allowed_values[:-1]])
-        options += f', and {repr(allowed_values[-1])}'
+        if len(allowed_values) == 2:
+            options += ' and '.join(repr(v) for v in allowed_values)
+        else:
+            options += ', '.join(repr(v) for v in allowed_values[:-1])
+            options += f', and {repr(allowed_values[-1])}'
     raise ValueError(msg.format(parameter=parameter, options=options,
                                 value=value, extra=extra))
 
