### 2021-11-03 81a87b1
MAINT: More deprecations and bumps
