+++ mne/connectivity/tests/test_envelope.py
@@ -1,100 +1,0 @@
-# Authors: Eric Larson <larson.eric.d@gmail.com>
-#          Sheraz Khan <sheraz@khansheraz.com>
-#          Denis Engemann <denis.engemann@gmail.com>
-#
-# License: BSD-3-Clause
-
-import numpy as np
-from numpy.testing import assert_allclose
-import pytest
-from scipy.signal import hilbert
-
-from mne.connectivity import envelope_correlation
-
-
-def _compute_corrs_orig(data):
-    # This is the version of the code by Sheraz and Denis.
-    # For this version (epochs, labels, time) must be -> (labels, time, epochs)
-    n_epochs, n_labels, _ = data.shape
-    corr = np.zeros((n_labels, n_labels))
-    for epoch_data in data:
-        for ii in range(n_labels):
-            for jj in range(n_labels):
-                # Get timeseries for each pair
-                x, y = epoch_data[ii], epoch_data[jj]
-                x_mag = np.abs(x)
-                x_conj_scaled = x.conj()
-                x_conj_scaled /= x_mag
-                # Calculate orthogonalization
-                y_orth_x = (y * x_conj_scaled).imag
-                y_orth_x_mag = np.abs(y_orth_x)
-                # Estimate correlation
-                corr[ii, jj] += np.abs(np.corrcoef(x_mag, y_orth_x_mag)[0, 1])
-    corr = (corr + corr.T) / (2. * n_epochs)
-    corr.flat[::n_labels + 1] = 0.
-    return corr
-
-
-def test_envelope_correlation():
-    """Test the envelope correlation function."""
-    rng = np.random.RandomState(0)
-    data = rng.randn(2, 4, 64)
-    data_hilbert = hilbert(data, axis=-1)
-    corr_orig = _compute_corrs_orig(data_hilbert)
-    assert (0 <= corr_orig).all()
-    assert (corr_orig < 1).all()
-    # using complex data
-    corr = envelope_correlation(data_hilbert)
-    assert_allclose(corr, corr_orig)
-    # using callable
-    corr = envelope_correlation(data_hilbert,
-                                combine=lambda data: np.mean(data, axis=0))
-    assert_allclose(corr, corr_orig)
-    # do Hilbert internally, and don't combine
-    corr = envelope_correlation(data, combine=None)
-    assert corr.shape == (data.shape[0],) + corr_orig.shape
-    corr = np.mean(corr, axis=0)
-    assert_allclose(corr, corr_orig)
-    # degenerate
-    with pytest.raises(ValueError, match='float'):
-        envelope_correlation(data.astype(int))
-    with pytest.raises(ValueError, match='entry in data must be 2D'):
-        envelope_correlation(data[np.newaxis])
-    with pytest.raises(ValueError, match='n_nodes mismatch'):
-        envelope_correlation([rng.randn(2, 8), rng.randn(3, 8)])
-    with pytest.raises(ValueError, match='mean or callable'):
-        envelope_correlation(data, 1.)
-    with pytest.raises(ValueError, match='Combine option'):
-        envelope_correlation(data, 'foo')
-    with pytest.raises(ValueError, match='Invalid value.*orthogonalize.*'):
-        envelope_correlation(data, orthogonalize='foo')
-
-    corr_plain = envelope_correlation(data, combine=None, orthogonalize=False)
-    assert corr_plain.shape == (data.shape[0],) + corr_orig.shape
-    assert np.min(corr_plain) < 0
-    corr_plain_mean = np.mean(corr_plain, axis=0)
-    assert_allclose(np.diag(corr_plain_mean), 1)
[…]
