+++ mne/time_frequency/csd.py
@@ -10,7 +10,6 @@ import numbers
 
 import numpy as np
 from .tfr import _cwt_array, morlet, _get_nfft
-from ..fixes import _import_fft
 from ..io.pick import pick_channels, _picks_to_idx
 from ..utils import (logger, verbose, warn, copy_function_doc_to_method_doc,
                      ProgressBar)
@@ -700,7 +699,7 @@ def csd_array_fourier(X, sfreq, t0=0, fmin=0, fmax=np.inf, tmin=None,
     csd_morlet
     csd_multitaper
     """
-    rfftfreq = _import_fft('rfftfreq')
+    from scipy.fft import rfftfreq
     X, times, tmin, tmax, fmin, fmax = _prepare_csd_array(
         X, sfreq, t0, tmin, tmax, fmin, fmax)
 
@@ -848,7 +847,7 @@ def csd_array_multitaper(X, sfreq, t0=0, fmin=0, fmax=np.inf, tmin=None,
     csd_morlet
     csd_multitaper
     """
-    rfftfreq = _import_fft('rfftfreq')
+    from scipy.fft import rfftfreq
     X, times, tmin, tmax, fmin, fmax = _prepare_csd_array(
         X, sfreq, t0, tmin, tmax, fmin, fmax)
 
