+++ mne/datasets/fnirs_motor/__init__.py
@@ -1,3 +1,3 @@
 """fNIRS motor dataset."""
 
-from .fnirs_motor import data_path, has_fnirs_motor_data, get_version
+from .fnirs_motor import data_path, get_version
