+++ mne/datasets/somato/__init__.py
@@ -1,3 +1,3 @@
 """Somatosensory dataset."""
 
-from .somato import data_path, has_somato_data, get_version
+from .somato import data_path, get_version
