### 2021-11-03 0af986c
MAINT: Use all conda-forge
