+++ mne/viz/_brain/tests/test_brain.py
@@ -384,8 +384,7 @@ def test_brain_init(renderer_pyvistaqt, tmp_path, pixel_ratio, brain_gc):
     assert_allclose(cam.GetRoll(), previous_roll + view_args["roll"])
     del view_args
 
-    with pytest.warns(DeprecationWarning, match='`view` is a dict'):
-        brain.show_view(view=dict(azimuth=180., elevation=90.))
+    brain.show_view(view=dict(azimuth=180., elevation=90.))
 
     # image and screenshot
     fname = op.join(str(tmp_path), 'test.png')
