+++ mne/io/meas_info.py
@@ -715,8 +715,7 @@ class Info(dict, MontageMixin):
         if key in self._attributes:
             if isinstance(self._attributes[key], str):
                 if not unlocked:
-                    warn(f'{self._attributes[key]} This warning will turn '
-                         'into an error after 0.24', DeprecationWarning)
+                    raise RuntimeError(self._attributes[key])
             else:
                 val = self._attributes[key](val)  # attribute checker function
         else:
