+++ mne/report/report.py
@@ -40,7 +40,7 @@ from ..proj import read_proj
 from .._freesurfer import _reorient_image, _mri_orientation
 from ..utils import (logger, verbose, get_subjects_dir, warn, _ensure_int,
                      fill_doc, _check_option, _validate_type, _safe_input,
-                     _path_like, use_log_level, deprecated, _check_fname,
+                     _path_like, use_log_level, _check_fname,
                      _check_ch_locs)
 from ..viz import (plot_events, plot_alignment, plot_cov, plot_projs_topomap,
                    plot_compare_evokeds, set_3d_view, get_3d_backend)
@@ -744,22 +744,6 @@ class Report(object):
     @property
     def html(self):
         return [element.html for element in self._content]
-
-    @property
-    def fnames(self):
-        warn(
-            message='Report.fnames is deprecated',
-            category=DeprecationWarning
-        )
-        return [element.name for element in self._content]
-
-    @property
-    def sections(self):
-        warn(
-            message='Report.sections is deprecated. Use Reports.tags instead',
-            category=DeprecationWarning
-        )
-        return self.tags
 
     @property
     def tags(self):
@@ -1650,8 +1634,7 @@ class Report(object):
             replace=replace
         )
 
-    def remove(self, caption=None, section=None, *, title=None, tags=None,
-               remove_all=False):
+    def remove(self, *, title=None, tags=None, remove_all=False):
         """Remove elements from the report.
 
         The element to remove is searched for by its title. Optionally, tags
@@ -1660,17 +1643,6 @@ class Report(object):
 
         Parameters
         ----------
-        caption : str
-            Remove content based on its caption.
-
-            .. deprecated:: 0.24.0
-               This parameter is scheduled for removal. Use ``title`` instead.
-        section : str | None
-            If supplied, restrict the operation to elements within the supplied
-            section.
-
-            .. deprecated:: 0.24.0
-               This parameter is scheduled for removal. Use ``tags`` instead.
         title : str
             The title of the element(s) to remove.
 
@@ -1698,21 +1670,6 @@ class Report(object):
             .. versionchanged:: 0.24.0
                Returns tuple if ``remove_all`` is ``True``.
         """
-        if caption is not None:
-            warn(
-                message='The "caption" parameter has been deprecated. Please '
-                        'use "title" instead',
-                category=DeprecationWarning
-            )
-            title = caption
-        if section is not None:
-            warn(
-                message='The "section" parameter has been deprecated. Please '
-                        'use "tags" instead',
-                category=DeprecationWarning
-            )
-            tags = _clean_tags(section)
-
         remove_idx = []
         for idx, element in enumerate(self._content):
             if element.name == title:
@@ -1928,98 +1885,6 @@ class Report(object):
             replace=replace
         )
 
-    @deprecated('Use :meth:`~mne.Report.add_figure` instead')
-    @fill_doc
-    def add_figs_to_section(self, figs, captions, section='custom',
-                            scale=None, image_format=None, comments=None,
-                            replace=False, auto_close=True):
-        """Append custom user-defined figures.
-
-        Parameters
-        ----------
-        figs : matplotlib.figure.Figure | mlab.Figure | array | list
-            A figure or a list of figures to add to the report. Each figure in
-            the list can be an instance of :class:`matplotlib.figure.Figure`,
-            :class:`mayavi.core.api.Scene`, or :class:`numpy.ndarray`.
-        captions : str | list of str
-            A caption or a list of captions to the figures.
-        section : str
-            Name of the section to place the figures in. If it already
-            exists, the figures will be appended to the end of the existing
-            section.
-        scale : float | None | callable
-            Scale the images maintaining the aspect ratio.
[…]
