+++ azure-pipelines.yml
@@ -81,7 +81,7 @@ stages:
   variables:
     AZURE_CI: 'true'
   jobs:
-  - job: Ultraslow+PG
+  - job: Ultraslow_PG
     pool:
       vmImage: 'ubuntu-20.04'
     variables:
