+++ mne/preprocessing/eog.py
@@ -8,7 +8,7 @@ import numpy as np
 
 from ._peak_finder import peak_finder
 from .. import pick_types, pick_channels
-from ..utils import logger, verbose, _pl, warn, _validate_type
+from ..utils import logger, verbose, _pl, _validate_type
 from ..filter import filter_data
 from ..epochs import Epochs
 
