+++ mne/time_frequency/tfr.py
@@ -26,13 +26,13 @@ from ..utils import (logger, verbose, _time_mask, _freq_mask, check_fname,
                      _gen_events, SizeMixin, _is_numeric, _check_option,
                      _validate_type, _check_combine, _check_pandas_installed,
                      _check_pandas_index_arguments, _check_time_format,
-                     _convert_times, _build_data_frame)
+                     _convert_times, _build_data_frame, warn)
 from ..channels.channels import ContainsMixin, UpdateChannelsMixin
 from ..channels.layout import _merge_ch_data, _pair_grad_sensors
 from ..io.pick import (pick_info, _picks_to_idx, channel_type, _pick_inst,
                        _get_channel_types)
 from ..io.meas_info import Info
-from ..viz.utils import (figure_nobar, plt_show, _setup_cmap, warn,
+from ..viz.utils import (figure_nobar, plt_show, _setup_cmap,
                          _connection_line, _prepare_joint_axes,
                          _setup_vmin_vmax, _set_title_multiple_electrodes)
 from ..externals.h5io import write_hdf5, read_hdf5
