### 2021-11-03 db1c29c
FIX: Error class
