### 2021-11-03 a7bbdf0
MRG, MAINT: Remove Mayavi (#9945)
* MAINT: Remove Mayavi

* FIX: Missed

* FIX: Flake

* FIX: Restore

* FIX: Commas

* FIX: Fix message

* FIX: subjects_dir

* FIX: Fix ref

* Better

* FIX: Azure

* FIX: Fix doc

* FIX: VTK

* FIX: Whoops
