+++ mne/conftest.py
@@ -80,7 +80,6 @@ def pytest_configure(config):
     ignore::ImportWarning
     ignore:the matrix subclass:PendingDeprecationWarning
     ignore:numpy.dtype size changed:RuntimeWarning
-    ignore:.*HasTraits.trait_.*:DeprecationWarning
     ignore:.*takes no parameters:DeprecationWarning
     ignore:joblib not installed:RuntimeWarning
     ignore:Using a non-tuple sequence for multidimensional indexing:FutureWarning
@@ -102,7 +101,6 @@ def pytest_configure(config):
     ignore:scipy\.gradient is deprecated.*:DeprecationWarning
     ignore:sklearn\.externals\.joblib is deprecated.*:FutureWarning
     ignore:The sklearn.*module.*deprecated.*:FutureWarning
-    ignore:.*trait.*handler.*deprecated.*:DeprecationWarning
     ignore:.*rich_compare.*metadata.*deprecated.*:DeprecationWarning
     ignore:.*In future, it will be an error for 'np.bool_'.*:DeprecationWarning
     ignore:.*`np.bool` is a deprecated alias.*:DeprecationWarning
@@ -118,8 +116,6 @@ def pytest_configure(config):
     ignore:Unable to enable faulthandler.*:UserWarning
     ignore:Fetchers from the nilearn.*:FutureWarning
     ignore:SelectableGroups dict interface is deprecated\. Use select\.:DeprecationWarning
-    ignore:Call to deprecated class vtk.*:DeprecationWarning
-    ignore:Call to deprecated method.*Deprecated since.*:DeprecationWarning
     always:.*get_data.* is deprecated in favor of.*:DeprecationWarning
     ignore:.*rcParams is deprecated.*global_theme.*:DeprecationWarning
     ignore:.*distutils\.sysconfig module is deprecated.*:DeprecationWarning
@@ -130,8 +126,6 @@ def pytest_configure(config):
     ignore:^Please use.*scipy\..*:DeprecationWarning
     ignore:.*Passing a schema to Validator.*:DeprecationWarning
     ignore:.*Found the following unknown channel type.*:RuntimeWarning
-    ignore:.*in an Any trait will be shared.*:DeprecationWarning
-    ignore:.*Mayavi 3D backend is deprecated.*:DeprecationWarning
     ignore:.*np\.MachAr.*:DeprecationWarning
     always::ResourceWarning
     """  # noqa: E501
@@ -202,15 +196,6 @@ def matplotlib_config():
     # functionality)
     plt.ioff()
     plt.rcParams['figure.dpi'] = 100
-    try:
-        from traits.etsconfig.api import ETSConfig
-    except Exception:
-        pass
-    else:
-        try:
-            ETSConfig.toolkit = 'qt4'
-        except Exception:
-            pass  # 'null' might be the only option in some configs
 
     # Make sure that we always reraise exceptions in handlers
     orig = cbook.CallbackRegistry
@@ -446,7 +431,7 @@ def browser_backend(request, garbage_collect):
         backend._close_all()
 
 
-@pytest.fixture(params=["mayavi", "pyvistaqt"])
+@pytest.fixture(params=["pyvistaqt"])
 def renderer(request, garbage_collect):
     """Yield the 3D backends."""
     with _use_backend(request.param, interactive=False) as renderer:
@@ -456,13 +441,6 @@ def renderer(request, garbage_collect):
 @pytest.fixture(params=["pyvistaqt"])
 def renderer_pyvistaqt(request, garbage_collect):
     """Yield the PyVista backend."""
-    with _use_backend(request.param, interactive=False) as renderer:
-        yield renderer
-
-
-@pytest.fixture(params=["mayavi"])
-def renderer_mayavi(request, garbage_collect):
-    """Yield the mayavi backend."""
[…]
