+++ mne/viz/_3d.py
@@ -20,7 +20,7 @@ from functools import partial
 import numpy as np
 
 from ..defaults import DEFAULTS
-from ..fixes import _crop_colorbar, _get_img_fdata, _get_args
+from ..fixes import _crop_colorbar, _get_img_fdata
 from .._freesurfer import (_read_mri_info, _check_mri, _get_head_surface,
                            _get_skull_surface)
 from ..io import _loc_to_coil_trans
@@ -318,7 +318,7 @@ def plot_evoked_field(evoked, surf_maps, time=None, time_label='t = %0.0f ms',
     time_label : str | None
         How to print info about the time instant visualized.
     %(n_jobs)s
-    fig : instance of mayavi.core.api.Scene | None
+    fig : instance of PyVista renderer | None
         If None (default), a new figure will be created, otherwise it will
         plot into the given figure.
 
@@ -335,8 +335,8 @@ def plot_evoked_field(evoked, surf_maps, time=None, time_label='t = %0.0f ms',
 
     Returns
     -------
-    fig : instance of mayavi.mlab.Figure
-        The mayavi figure.
+    fig : instance of PyVista renderer
+        The figure.
     """
     # Update the backend
     from .backends.renderer import _get_renderer
@@ -356,7 +356,7 @@ def plot_evoked_field(evoked, surf_maps, time=None, time_label='t = %0.0f ms',
     # Plot them
     alphas = [1.0, 0.5]
     colors = [(0.6, 0.6, 0.6), (1.0, 1.0, 1.0)]
-    colormap = mne_analyze_colormap(format='mayavi')
+    colormap = mne_analyze_colormap(format='vtk')
     colormap_lines = np.concatenate([np.tile([0., 0., 255., 255.], (127, 1)),
                                      np.tile([0., 0., 0., 255.], (2, 1)),
                                      np.tile([255., 0., 0., 255.], (127, 1))])
@@ -493,8 +493,8 @@ def plot_alignment(info=None, trans=None, subject=None, subjects_dir=None,
 
         .. versionadded:: 0.16
     %(dbs)s
-    fig : mayavi.mlab.Figure | None
-        Mayavi Scene in which to plot the alignment.
+    fig : PyVista renderer | None
+        PyVista scene in which to plot the alignment.
         If ``None``, creates a new 600x600 pixel figure with black background.
 
         .. versionadded:: 0.16
@@ -507,8 +507,8 @@ def plot_alignment(info=None, trans=None, subject=None, subjects_dir=None,
 
     Returns
     -------
-    fig : instance of mayavi.mlab.Figure
-        The mayavi figure.
+    fig : instance of PyVista renderer
+        The figure.
 
     See Also
     --------
@@ -1704,10 +1704,6 @@ def _check_volume(stc, src, surface, backend_name):
     if isinstance(stc, _BaseSurfaceSourceEstimate):
         return False
     else:
-        if backend_name == 'mayavi':
-            raise RuntimeError(
-                'Must use the PyVista 3D backend to plot a mixed or volume '
-                'source estimate')
         _validate_type(src, SourceSpaces, 'src',
                        'src when stc is a mixed or volume source estimate')
         if isinstance(stc, _BaseMixedSourceEstimate):
@@ -1764,10 +1760,10 @@ def plot_source_estimates(stc, subject=None, surface='inflated', hemi='lh',
         .. versionchanged:: 0.20.0
            "auto" mode added.
     %(subjects_dir)s
-    figure : instance of mayavi.core.api.Scene | instance of matplotlib.figure.Figure | list | int | None
+    figure : instance of PyVista renderer | instance of matplotlib.figure.Figure | list | int | None
         If None, a new figure will be created. If multiple views or a
         split view is requested, this must be a list of the appropriate
-        length. If int is provided it will be used to identify the Mayavi
[…]
