### 2021-11-03 5d76216
better test for _voxel_neighbors
