### 2021-11-04 e9126c8
add local maxima button to ieeg locate gui
