### 2021-11-04 7892c08
MRG, MAINT: Last deps (#9949)
* MAINT: Last deps

* FIX: Flake

* FIX: Error class

* FIX: Missed one

* FIX: Deps
