### 2021-11-04 c69ec3d
MRG, MAINT: More deprecations and bumps (#9947)
* MAINT: More deprecations and bumps

* FIX: Imports

* FIX: Doc

* FIX: One more

* FIX: Azure

* FIX: Reqs

* FIX: Missed a test
