+++ mne/datasets/ssvep/ssvep.py
@@ -1,18 +1,9 @@
 # Authors: Dominik Welke <dominik.welke@web.de>
 # License: BSD Style.
 
-from ...utils import verbose, deprecated
-from ..utils import (has_dataset, _data_path_doc, DEPRECATION_MESSAGE_TEMPLATE,
-                     _get_version, _version_doc, _download_mne_dataset,
-                     _HAS_DATA_DOCSTRING_TEMPLATE)
-
-
-@deprecated(extra=DEPRECATION_MESSAGE_TEMPLATE.format('ssvep'))
-def has_ssvep_data():
-    return has_dataset(name='ssvep')
-
-
-has_ssvep_data.__doc__ = _HAS_DATA_DOCSTRING_TEMPLATE.format('ssvep')
+from ...utils import verbose
+from ..utils import (_data_path_doc, _get_version, _version_doc,
+                     _download_mne_dataset)
 
 
 @verbose
