### 2021-11-04 d6cf179
MAINT: Add config var for opt-in [ci skip]
