+++ mne/utils/config.py
@@ -70,6 +70,7 @@ known_config_types = (
     'MNE_3D_OPTION_ANTIALIAS',
     'MNE_BROWSE_RAW_SIZE',
     'MNE_BROWSE_BACKEND',
+    'MNE_BROWSE_USE_OPENGL',
     'MNE_CACHE_DIR',
     'MNE_COREG_ADVANCED_RENDERING',
     'MNE_COREG_COPY_ANNOT',
