+++ mne/utils/docs.py
@@ -1529,11 +1529,12 @@ precompute : bool | str
 """
 
 docdict['use_opengl'] = """
-use_opengl : bool
+use_opengl : bool | None
     Whether to use OpenGL when rendering the plot (requires ``pyopengl``).
     May increase performance, but effect is dependent on system CPU and
-    graphics hardware. Only works if using the pyQtGraph backend. Default is
-    ``True``.
+    graphics hardware. Only works if using the PyQtGraph backend. Default is
+    None, which will use False unless the user configuration variable
+    ``MNE_BROWSE_USE_OPENGL`` is set to ``'true'``, see :func:`mne.set_config`.
 
     .. versionadded:: 0.24
 """
