+++ mne/viz/raw.py
@@ -36,7 +36,7 @@ def plot_raw(raw, events=None, duration=10.0, start=0.0, n_channels=20,
              proj=True, group_by='type', butterfly=False, decim='auto',
              noise_cov=None, event_id=None, show_scrollbars=True,
              show_scalebars=True, time_format='float',
-             precompute='auto', use_opengl=True, verbose=None):
+             precompute='auto', use_opengl=None, verbose=None):
     """Plot raw data.
 
     Parameters
@@ -316,6 +316,7 @@ def plot_raw(raw, events=None, duration=10.0, start=0.0, n_channels=20,
         raise TypeError(f'title must be None or a string, got a {type(title)}')
 
     # gather parameters and initialize figure
+    _validate_type(use_opengl, (bool, None), 'use_opengl')
     params = dict(inst=raw,
                   info=info,
                   # channels and channel order
