### 2021-11-04 cbd0387
Update mne/gui/tests/test_ieeg_locate_gui.py
Co-authored-by: Eric Larson <larson.eric.d@gmail.com>
