+++ mne/viz/__init__.py
@@ -11,8 +11,7 @@ from ._3d import (plot_sparse_source_estimates, plot_source_estimates,
                   plot_vector_source_estimates, plot_evoked_field,
                   plot_dipole_locations, snapshot_brain_montage,
                   plot_head_positions, plot_alignment, plot_brain_colorbar,
-                  plot_volume_source_estimates, plot_sensors_connectivity,
-                  link_brains, set_3d_options)
+                  plot_volume_source_estimates, link_brains, set_3d_options)
 from .misc import (plot_cov, plot_csd, plot_bem, plot_events,
                    plot_source_spectrogram, _get_presser,
                    plot_dipole_amplitudes, plot_ideal_filter, plot_filter,
@@ -20,8 +19,7 @@ from .misc import (plot_cov, plot_csd, plot_bem, plot_events,
 from .evoked import (plot_evoked, plot_evoked_image, plot_evoked_white,
                      plot_snr_estimate, plot_evoked_topo,
                      plot_evoked_joint, plot_compare_evokeds)
-from .circle import (plot_connectivity_circle, circular_layout,
-                     plot_channel_labels_circle)
+from .circle import (circular_layout, plot_channel_labels_circle)
 from .epochs import (plot_drop_log, plot_epochs, plot_epochs_psd,
                      plot_epochs_image)
 from .raw import plot_raw, plot_raw_psd, plot_raw_psd_topo, _RAW_CLIP_DEF
