### 2021-11-04 4d15103
[MAINT] Remove viz connectivity functions (#9959)
* Remove files from conn

* Fix import statements

* Fix api docs

* Fix docstring

* Fix changelog
