+++ doc/visualization.rst
@@ -56,7 +56,6 @@ Visualization
    plot_raw
    plot_raw_psd
    plot_sensors
-   plot_sensors_connectivity
    plot_snr_estimate
    plot_source_estimates
    link_brains
