### 2021-11-05 55f5491
Add 'maxshield' to docstring.
