### 2021-11-05 78c925c
Add 'mri_head_t' to docstring.
