### 2021-11-05 d924754
Add 'kit_system_id' to docstring.
