### 2021-11-05 bbd21a5
Added 'meas_file' and 'mri_file' to docstring.
