+++ mne/io/meas_info.py
@@ -358,9 +358,13 @@ class Info(dict, MontageMixin):
         .. versionchanged:: 0.20
            This is stored as a :class:`~python:datetime.datetime` object
            instead of a tuple of seconds/microseconds.
+    meas_file : str | None
+        Measurement file (used for source estimation).
     meas_id : dict | None
         The ID assigned to this measurement by the acquisition system or
         during file conversion. Follows the same format as ``file_id``.
+    mri_file : str | None
+        MRI file (used for source estimation).
     mri_head_t : dict | None
         Transformation from MRI to head coordinates.
     nchan : int
@@ -656,9 +660,9 @@ class Info(dict, MontageMixin):
         'maxshield': 'maxshield cannot be set directly.',
         'meas_date': 'meas_date cannot be set directly. '
                      'Please use method inst.set_meas_date() instead.',
-        'meas_file': 'meas_file cannot be set directly.',  # doc?
+        'meas_file': 'meas_file cannot be set directly.',
         'meas_id': 'meas_id cannot be set directly.',
-        'mri_file': 'mri_file cannot be set directly.',  # doc?
+        'mri_file': 'mri_file cannot be set directly.',
         'mri_head_t': 'mri_head_t cannot be set directly.',
         'mri_id': 'mri_id cannot be set directly.',  # doc?
         'nchan': 'nchan cannot be set directly. '
