### 2021-11-05 401b52a
Fix docstring for 'meas_file' and 'mri_file'.
