### 2021-11-05 d10e25f
Remove outdated comment.
