### 2021-11-05 c2ee7ae
[MAINT, MRG] Better test for `mne.surface._voxel_neighbors` (#9950)
* better test for _voxel_neighbors

* bad import
