### 2021-11-05 316e429
Add all attributes that can be modified to the docstring. [skip azp] [skip actions]
