### 2021-11-05 c2aaaa8
[ENH, MRG] Add local maxima button to ieeg locate GUI (#9952)
* add local maxima button to ieeg locate gui

* add call to draw

* update latest PR number

* Update mne/gui/tests/test_ieeg_locate_gui.py

Co-authored-by: Eric Larson <larson.eric.d@gmail.com>

Co-authored-by: Eric Larson <larson.eric.d@gmail.com>
