### 2021-11-05 3ff0c4c
Remove check_consistency on key 'filename' and tests on 'filename' as this is now covered by locking/unlocking.
