### 2021-11-05 88b63a8
Remove filename from list of attributes.
