### 2021-11-05 105c597
Add 'command_line' and 'working_dir' to the docstring.
