### 2021-11-05 17d2e38
Add 'xplotter_layout' to the docstring.
