+++ mne/io/meas_info.py
@@ -401,9 +401,12 @@ class Info(dict, MontageMixin):
         "UTC offset of related meas_date (sHH:MM).
 
         .. versionadded:: 0.19
+
     working_dir : str
         Working directory used when the source space was created (used for
         source estimation).
+    xplotter_layout : str
+        Layout of the Xplotter (Neuromag system only).
 
     See Also
     --------
@@ -650,7 +653,7 @@ class Info(dict, MontageMixin):
                'inst.drop_channels(), inst.pick_channels(), '
                'inst.rename_channels(), inst.reorder_channels() '
                'and inst.set_channel_types() instead.',
-        'command_line': 'command_line cannot be set directly.',  # doc?
+        'command_line': 'command_line cannot be set directly.',
         'comps': 'comps cannot be set directly. '
                  'Please use method Raw.apply_gradient_compensation() '
                  'instead.',
@@ -700,8 +703,8 @@ class Info(dict, MontageMixin):
         'subject_info': _check_subject_info,
         'temp': lambda x: x,
         'utc_offset': 'utc_offset cannot be set directly.',
-        'working_dir': 'working_dir cannot be set directly.',  # doc?
-        'xplotter_layout': 'xplotter_layout cannot be set directly.'  # doc?
+        'working_dir': 'working_dir cannot be set directly.',
+        'xplotter_layout': 'xplotter_layout cannot be set directly.'
     }
 
     def __init__(self, *args, **kwargs):
