### 2021-11-05 8d3f48b
MRG, BUG: Fix regression with report due to bbox_inches (#9966)
* BUG: Fix regression with report due to bbox_inches

* FIX: Link
