### 2021-11-05 8efb01f
MRG, BUG: Fix regression with report due to bbox_inches (#9966)
* BUG: Fix regression with report due to bbox_inches

* FIX: Link
