+++ azure-pipelines.yml
@@ -245,7 +245,7 @@ stages:
     - script: conda env update --name base --file environment.yml
       condition: eq(variables['TEST_MODE'], 'conda')
       displayName: Setup MNE environment
-    - script: conda remove -c conda-forge -yq mne
+    - script: conda remove -c conda-forge --force -yq mne
       condition: eq(variables['TEST_MODE'], 'conda')
       displayName: Remove old MNE
     - script: pip install -e .
@@ -305,7 +305,7 @@ stages:
       displayName: Setup MNE environment
     - script: conda install -c conda-forge graphviz
       displayName: Add graphviz dot
-    - script: conda remove -c conda-forge -yq mne
+    - script: conda remove -c conda-forge --force -yq mne
       displayName: Remove old MNE
     - script: pip install -e .
       displayName: Install dev MNE
