+++ azure-pipelines.yml
@@ -248,7 +248,7 @@ stages:
     - bash: |
         set -e
         conda remove -c conda-forge --force -yq mne
-        rm /Miniconda/Scripts/mne.exe
+        rm /c/Miniconda/Scripts/mne.exe
       condition: eq(variables['TEST_MODE'], 'conda')
       displayName: Remove old MNE
     - script: pip install -e .
@@ -311,8 +311,6 @@ stages:
     - bash: |
         set -e
         conda remove -c conda-forge --force -yq mne
-        echo $PATH
-        which mne
         rm /c/Miniconda/Scripts/mne.exe
       displayName: Remove old MNE
     - script: pip install -e .
