### 2021-11-05 e62543c
Add extensive checking for Report tags
Fixes #9965
