+++ mne/report/report.py
@@ -10,6 +10,7 @@ import io
 import dataclasses
 from dataclasses import dataclass
 from typing import Tuple
+from collections.abc import Sequence
 import base64
 from io import BytesIO, StringIO
 import contextlib
@@ -276,6 +277,41 @@ class _ContentElement:
     dom_id: str
     tags: Tuple[str]
     html: str
+
+
+def _check_tags(tags) -> Tuple[str]:
+    # Must be iterable, not not a string
+    if (isinstance(tags, str) or not isinstance(tags, (Sequence, np.ndarray))):
+        raise TypeError(
+            f'tags must be a collection of str, but got {type(tags)} '
+            f'instead: {tags}'
+        )
+    tags = tuple(tags)
+
+    # Check for invalid dtypes
+    bad_tags = [tag for tag in tags
+                if not isinstance(tag, str)]
+    if bad_tags:
+        raise TypeError(
+            f'tags must be strings, but got the following instead: '
+            f'{", ".join([str(tag) for tag in bad_tags])}'
+        )
+
+    # Check for invalid characters
+    invalid_chars = (' ', '"', '\n')  # we'll probably find more :-)
+    bad_tags = []
+    for tag in tags:
+        for invalid_char in invalid_chars:
+            if invalid_char in tag:
+                bad_tags.append(tag)
+                break
+    if bad_tags:
+        raise ValueError(
+            f'The following tags contained invalid characters: '
+            f'{", ".join(bad_tags)}'
+        )
+
+    return tags
 
 
 ###############################################################################
@@ -807,7 +843,7 @@ class Report(object):
         -----
         .. versionadded:: 0.24.0
         """
-        tags = tuple(tags)
+        tags = _check_tags(tags)
 
         add_projs = self.projs if projs is None else projs
 
@@ -903,7 +939,7 @@ class Report(object):
             noise_cov = self.cov_fname
         if noise_cov is not None and not isinstance(noise_cov, Covariance):
             noise_cov = read_cov(fname=noise_cov)
-        tags = tuple(tags)
+        tags = _check_tags(tags)
 
         add_projs = self.projs if projs is None else projs
 
@@ -968,7 +1004,7 @@ class Report(object):
         -----
         .. versionadded:: 0.24.0
         """
-        tags = tuple(tags)
+        tags = _check_tags(tags)
 
         if psd is None:
             add_psd = dict() if self.raw_psd is True else self.raw_psd
@@ -1038,7 +1074,7 @@ class Report(object):
         -----
         .. versionadded:: 0.24.0
         """
-        tags = tuple(tags)
+        tags = _check_tags(tags)
 
         html, dom_id = self._render_stc(
             stc=stc,
@@ -1083,7 +1119,7 @@ class Report(object):
         -----
         .. versionadded:: 0.24.0
         """
-        tags = tuple(tags)
+        tags = _check_tags(tags)
 
         html, dom_id = self._render_forward(
             forward=forward, subject=subject, subjects_dir=subjects_dir,
@@ -1126,7 +1162,7 @@ class Report(object):
         -----
         .. versionadded:: 0.24.0
         """
-        tags = tuple(tags)
+        tags = _check_tags(tags)
 
         if ((subject is not None and trans is None) or
                 (trans is not None and subject is None)):
@@ -1172,7 +1208,7 @@ class Report(object):
         -----
         .. versionadded:: 0.24.0
         """
-        tags = tuple(tags)
+        tags = _check_tags(tags)
 
         html, dom_id = self._render_trans(
             trans=trans,
@@ -1210,7 +1246,7 @@ class Report(object):
         -----
         .. versionadded:: 0.24.0
         """
-        tags = tuple(tags)
[…]
