### 2021-11-05 6a1f726
Update mne/report/report.py
