### 2021-11-06 ffa8ae8
Add "interaction" param to CoregistrationUI
