### 2021-11-07 c8d3037
Fix docstring
