### 2021-11-07 #3897 @prashjha
Thank you @TomDonoghue and @janfreyberg for agreeing to be reviewers. Please read the first couple of comments in this thread and also reviewer guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Also, you can browse the closed "REVIEW" issues on the "joss-reviews" repository to get some ideas on how to complete the reviews. Good luck!
