### 2021-11-07 #3897 @janfreyberg
Hi @alexrockhill, I can't seem to find obvious documentation for these features in the MNE docs, and the paper currently doesn't provide any information on what part of the library you are describing (in code terms - i.e. what module / submodule / class).

Could you please add this to the paper, and also include examples that demonstrate how the library is used? It's difficult to know what to assess otherwise, as MNE is so big!

Thanks!
