### 2021-11-07 #3897 @alexrockhill
The code described in this paper is for based in these three tutorials:

https://mne.tools/dev/auto_tutorials/clinical/10_ieeg_localize.html
https://mne.tools/dev/auto_tutorials/clinical/20_seeg.html
https://mne.tools/dev/auto_tutorials/clinical/30_ecog.html

and all the modules that they use, especially `mne.gui.locate_ieeg` and `mne.viz.Brain` (for visualizations).

I'll add that to the manuscript, thanks for the request for clarification!
