### 2021-11-08 6a72700
Fix for plot_alignment
