+++ mne/viz/_3d.py
@@ -695,7 +695,7 @@ def plot_alignment(info=None, trans=None, subject=None, subjects_dir=None,
     # initialize figure
     renderer = _get_renderer(fig, name=f'Sensor alignment: {subject}',
                              bgcolor=(0.5, 0.5, 0.5), size=(800, 800))
-    renderer.set_interaction('terrain')
+    renderer.set_interaction(interaction)
 
     # plot head
     _, _, head_surf = _plot_head_surface(
