### 2021-11-08 8ee5952
MRG, FIX: Fix interior check (#9968)
* FIX: Fix interior check

* FIX: Copy-paste error
