### 2021-11-08 83d25a3
MRG: Add "interaction" param to CoregistrationUI (#9972)
