+++ mne/export/tests/test_export.py
@@ -83,7 +83,7 @@ def test_double_export_edf(tmp_path):
 
     # export again
     raw_read.load_data()
-    raw_read.export(temp_fname, add_ch_type=True)
+    raw_read.export(temp_fname, add_ch_type=True, overwrite=True)
     raw_read = read_raw_edf(temp_fname, preload=True)
 
     # stim channel should be dropped
@@ -185,28 +185,28 @@ def test_rawarray_edf(tmp_path):
     raw_bad.rename_channels({'1': 'abcdefghijklmnopqrstuvwxyz'})
     with pytest.raises(RuntimeError, match='Signal label'), \
             pytest.warns(RuntimeWarning, match='Data has a non-integer'):
-        raw_bad.export(temp_fname)
+        raw_bad.export(temp_fname, overwrite=True)
 
     # include bad birthday that is non-EDF compliant
     bad_info = info.copy()
     bad_info['subject_info']['birthday'] = (1700, 1, 20)
     raw = RawArray(data, bad_info)
     with pytest.raises(RuntimeError, match='Setting patient birth date'):
-        raw.export(temp_fname)
+        raw.export(temp_fname, overwrite=True)
 
     # include bad measurement date that is non-EDF compliant
     raw = RawArray(data, info)
     meas_date = datetime(year=1984, month=1, day=1, tzinfo=timezone.utc)
     raw.set_meas_date(meas_date)
     with pytest.raises(RuntimeError, match='Setting start date time'):
-        raw.export(temp_fname)
+        raw.export(temp_fname, overwrite=True)
 
     # test that warning is raised if there are non-voltage based channels
     raw = RawArray(data, info)
     with pytest.warns(RuntimeWarning, match='The unit'):
         raw.set_channel_types({'9': 'hbr'})
     with pytest.warns(RuntimeWarning, match='Non-voltage channels'):
-        raw.export(temp_fname)
+        raw.export(temp_fname, overwrite=True)
 
     # data should match up to the non-accepted channel
     raw_read = read_raw_edf(temp_fname, preload=True)
