### 2021-11-08 74f8870
add manual alignment procedure to ieeg localize tutorial [skip CI] [skip circle]
