### 2021-11-08 184b747
Change versionadded from 1.0 to 0.24.1 [skip azp] [skip actions]
