### 2021-11-08 0d5f3ba
Add first samp and raw info sfreq
