+++ mne/epochs.py
@@ -2167,72 +2167,6 @@ class BaseEpochs(ProjMixin, ContainsMixin, UpdateChannelsMixin, ShiftTimeMixin,
         from .forward import _as_meg_type_inst
         return _as_meg_type_inst(self, ch_type=ch_type, mode=mode)
 
-    @property
-    def annotations(self):  # noqa: D102
-        return self._annotations
-
-    @verbose
-    def set_annotations(self, annotations, emit_warning=True,
-                        on_missing='raise', *, verbose=None):
-        """Setter for annotations.
-
-        This setter checks if they are inside the data range.
-
-        Parameters
-        ----------
-        annotations : instance of mne.Annotations | None
-            Annotations to set. If None, the annotations is defined
-            but empty.
-        emit_warning : bool
-            Whether to emit warnings when cropping or omitting annotations.
-        %(on_missing_ch_names)s
-        %(verbose_meth)s
-
-        Returns
-        -------
-        self : instance of Raw
-            The raw object with annotations.
-        """
-        meas_date = _handle_meas_date(self.info['meas_date'])
-        if annotations is None:
-            self._annotations = Annotations([], [], [], meas_date)
-        else:
-            _validate_type(annotations, Annotations, 'annotations')
-
-            if meas_date is None and annotations.orig_time is not None:
-                raise RuntimeError('Ambiguous operation. Setting an Annotation'
-                                   ' object with known ``orig_time`` to a raw'
-                                   ' object which has ``meas_date`` set to'
-                                   ' None is ambiguous. Please, either set a'
-                                   ' meaningful ``meas_date`` to the raw'
-                                   ' object; or set ``orig_time`` to None in'
-                                   ' which case the annotation onsets would be'
-                                   ' taken in reference to the first sample of'
-                                   ' the raw object.')
-
-            delta = 1. / self.info['sfreq']
-            new_annotations = annotations.copy()
-            new_annotations._prune_ch_names(self.info, on_missing)
-
-            # get the first time in continuous time in Raw
-
-            if annotations.orig_time is None:
-                # new_annotations.crop(0, self.times[-1] + delta,
-                #                      emit_warning=emit_warning)
-                new_annotations.onset += self._first_time
-            else:
-                tmin = meas_date + timedelta(0, self._first_time)
-                tmax = tmin + timedelta(seconds=self.times[-1] + delta)
-                # new_annotations.crop(tmin=tmin, tmax=tmax,
-                #                      emit_warning=emit_warning)
-                new_annotations.onset -= (
-                    meas_date - new_annotations.orig_time).total_seconds()
-            new_annotations._orig_time = meas_date
-
-            self._annotations = new_annotations
-
-        return self
-
 
 def _drop_log_stats(drop_log, ignore=('IGNORED',)):
     """Compute drop log stats.
@@ -2667,6 +2601,11 @@ class Epochs(BaseEpochs):
         proj = proj or raw.proj
 
         self.reject_by_annotation = reject_by_annotation
+
+        # store certain raw properties for dealing with annotations
+        self._first_time = raw._first_time.copy()
+        self._raw_sfreq = raw.info['sfreq']
+
         # call BaseEpochs constructor
         super(Epochs, self).__init__(
             info, None, events, event_id, tmin, tmax, metadata=metadata,
@@ -2719,6 +2658,72 @@ class Epochs(BaseEpochs):
                                             reject_start, reject_stop,
                                             self.reject_by_annotation)
         return data
+
+    @property
+    def annotations(self):  # noqa: D102
+        return self._annotations
+
+    @verbose
+    def set_annotations(self, annotations, emit_warning=True,
+                        on_missing='raise', *, verbose=None):
+        """Setter for annotations.
+
+        This setter checks if they are inside the data range.
+
+        Parameters
[…]
