+++ mne/io/eeglab/eeglab.py
@@ -81,7 +81,7 @@ def _to_loc(ll):
 
 def _eeg_has_montage_information(eeg):
     try:
-        from scipy.io._matlab import mat_struct
+        from scipy.io._matlab.mio5_params import mat_struct
     except ImportError:  # SciPy < 1.8
         from scipy.io.matlab.mio5_params import mat_struct
     if not len(eeg.chanlocs):
