### 2021-11-08 f863824
MNT: Rename browser-related config-values for consistency
