### 2021-11-08 6c758c6
revert renaming of MNE_BROWSE_RAW_SIZE
