+++ mne/utils/docs.py
@@ -1536,7 +1536,7 @@ use_opengl : bool | None
     May increase performance, but effect is dependent on system CPU and
     graphics hardware. Only works if using the PyQtGraph backend. Default is
     None, which will use False unless the user configuration variable
-    ``MNE_BROWSER_USE_OPENGL`` is set to ``'true'``, 
+    ``MNE_BROWSER_USE_OPENGL`` is set to ``'true'``,
     see :func:`mne.set_config`.
 
     .. versionadded:: 0.24
