### 2021-11-08 09621ea
import fix for scipy 1.8 pre (#9973)
* import fix for scipy 1.8 pre

* fix arg oversight

* FIX: A couple more

* FIX: Found another

* FIX: More [skip azp] [skip circle]

* FIX: Correct [skip circle] [skip azp]

* FIX: One more

Co-authored-by: Eric Larson <larson.eric.d@gmail.com>
