+++ mne/source_estimate.py
@@ -3352,21 +3352,20 @@ def stc_near_sensors(evoked, trans, subject, distance=0.01, mode='sum',
     subject = _check_subject(None, subject, raise_error=False)
     subjects_dir = get_subjects_dir(subjects_dir, raise_error=True)
     if surface is not None:
-        surf_rr = np.concatenate([
-            read_surface(op.join(subjects_dir, subject, 'surf',
-                                 f'{hemi}.{surface}'))[0]
-            for hemi in ('lh', 'rh')])
+        surf_rr = [read_surface(op.join(subjects_dir, subject, 'surf',
+                                        f'{hemi}.{surface}'))[0]
+                   for hemi in ('lh', 'rh')]
     if src is None:  # fake a full surface one
-        _validate_type(surface, str, 'surface', extra='when src is None')
-        rrs = surf_rr
+        _validate_type(surface, str, 'surface', 'when src is None')
         src = SourceSpaces([
             dict(rr=rr / 1000., vertno=np.arange(len(rr)), type='surf',
                  coord_frame=FIFF.FIFFV_COORD_MRI)
-            for rr in rrs])
+            for rr in surf_rr])
+        rrs = np.concatenate(surf_rr)
         keep_all = False
     else:
         if surface is not None:
-            rrs = surf_rr
+            rrs = np.concatenate(surf_rr)
         else:
             rrs = np.concatenate([s['rr'][s['vertno']] for s in src])
             if src[0]['coord_frame'] == FIFF.FIFFV_COORD_HEAD:
