### 2021-11-09 bd0f435
FIX: replace else statement with finally.
