### 2021-11-09 040d96b
Use shutil.rmtree instead of os.remove since mffpy creates a directory.
