+++ mne/export/_egimff.py
@@ -62,8 +62,7 @@ def export_evokeds_mff(fname, evoked, history=None, overwrite=False, *,
             shutil.rmtree(fname)
         else:
             raise
-    finally:
-        writer = mffpy.Writer(fname)
+    writer = mffpy.Writer(fname)
     current_time = pytz.utc.localize(datetime.datetime.utcnow())
     writer.addxml('fileInfo', recordTime=current_time)
     try:
