### 2021-11-09 51dbfed
improve comment preceding folder deletion.
