### 2021-11-09 f7dd7a8
Update mne/epochs.py [skip ci]
Co-authored-by: Eric Larson <larson.eric.d@gmail.com>
