### 2021-11-09 100a833
Update mne/export/_egimff.py [skip ci]
Co-authored-by: Eric Larson <larson.eric.d@gmail.com>
