### 2021-11-09 bbf8db5
Update mne/export/_export.py [skip ci]
Co-authored-by: Eric Larson <larson.eric.d@gmail.com>
