### 2021-11-09 5b36c81
Update mne/export/_export.py [skip ci]
Co-authored-by: Eric Larson <larson.eric.d@gmail.com>
