### 2021-11-09 7ba6930
Update mne/export/_egimff.py
Co-authored-by: Eric Larson <larson.eric.d@gmail.com>
