### 2021-11-09 708d258
Fix behavior when pathlib.Path are passed instead of strings.
