+++ mne/export/_egimff.py
@@ -3,8 +3,10 @@
 #
 # License: BSD-3-Clause
 
+import os
 import shutil
 import datetime
+import os.path as op
 
 import numpy as np
 
