### 2021-11-09 c9e5ddf
add test for epochs as well.
