### 2021-11-09 221e280
Remove str(fname) as it's not needed.
