+++ mne/io/base.py
@@ -1076,7 +1076,6 @@ class BaseRaw(ProjMixin, ContainsMixin, UpdateChannelsMixin, SetChannelsMixin,
         %(fir_window)s
         %(fir_design)s
         %(pad-fir)s
-            The default is ``'reflect_limited'``.
 
             .. versionadded:: 0.15
         %(verbose_meth)s
@@ -1160,7 +1159,7 @@ class BaseRaw(ProjMixin, ContainsMixin, UpdateChannelsMixin, SetChannelsMixin,
             An optional event matrix. When specified, the onsets of the events
             are resampled jointly with the data. NB: The input events are not
             modified, but a new array is returned with the raw instead.
-        %(pad-fir)s
+        %(pad)s
             The default is ``'reflect_limited'``.
 
             .. versionadded:: 0.15
