### 2021-11-09 396d11a
Add default in docstring.
