### 2021-11-09 7169435
MRG: Add argument overwrite to export for raws and epochs to match save. (#9975)
* Add argument overwrite to export for Raws.

* Add argument overwrite to export for epochs.

* Add tests.

* Add entry to changelog.

* Move entry to changelog to Bugs and fix spelling typo.

* fix indentation for ..versionadded

* typo fix.

* fix tests.

* Change versionadded from 1.0 to 0.24.1 [skip azp] [skip actions]

* Add overwrite for evoked instances (mffpy related).

* FIX: replace else statement with finally.

* Use shutil.rmtree instead of os.remove since mffpy creates a directory.

* Remove finally.

[…]
