### 2021-11-09 43bef66
Fix #9983 - Typo in docstring of pad for resampling methods. (#9984)
* Add pad-only docstring and build pad-fir docstring from pad-only.

* Change docdict entry for pad-fir to pad for resampling. [skip actions] [skip azp]

* Improve syntax [skip actions] [skip azp]

* Add default in docstring.

* revert [ci skip]

* fix [skip actions] [skip azp]

* Remove default in docstring for method filter as raw uses a different default value and the docstring is copied from FilterMixin [skip actions] [skip azp]
