+++ mne/filter.py
@@ -1122,6 +1122,7 @@ def notch_filter(x, Fs, freqs, filter_length='auto', notch_widths=None,
     %(fir_window)s
     %(fir_design)s
     %(pad-fir)s
+        The default is ``'reflect_limited'``.
     %(verbose)s
 
     Returns
@@ -1402,7 +1403,7 @@ def resample(x, up=1., down=1., npad=100, axis=-1, window='boxcar', n_jobs=1,
         Axis along which to resample (default is the last axis).
     %(window-resample)s
     %(n_jobs-cuda)s
-    %(pad-fir)s
+    %(pad)s
         The default is ``'reflect_limited'``.
 
         .. versionadded:: 0.15
@@ -2030,7 +2031,7 @@ class FilterMixin(object):
         %(npad)s
         %(window-resample)s
         %(n_jobs-cuda)s
-        %(pad-fir)s
+        %(pad)s
             The default is ``'edge'``, which pads with the edge values of each
             vector.
 
