### 2021-11-09 8924640
Adding unit tests for epochs
