### 2021-11-10 ed37d44
Make check_version() use _compare_version() internally
Get rid of `LooseVersion` usage here.
