### 2021-11-10 f7e9935
Add support to retrieve channel type from ['chanlocs']['type'].
