### 2021-11-10 7df21fa
Add small test of _get_montage_info
