### 2021-11-10 2ff40ff
Update paper.md [skip ci]
Co-authored-by: Daniel McCloy <dan@mccloy.info>
