### 2021-11-10 aa6ad2f
[ENH, MRG] Add function references [ci skip] (#9986)
* add specific function references [skip actions] [skip azp]

* Update paper.md [skip ci]

Co-authored-by: Daniel McCloy <dan@mccloy.info>

Co-authored-by: Daniel McCloy <dan@mccloy.info>
