### 2021-11-10 5c5c542
Update mne/io/eeglab/eeglab.py [ci skip]
Tested locally.

Co-authored-by: Eric Larson <larson.eric.d@gmail.com>
