### 2021-11-10 9aaffb5
Add warning and test.
