### 2021-11-10 0eaa085
Add more detailed warning.
