### 2021-12-11 #3897 @TomDonoghue
Hey all,

Sorry my review is a bit late - I had a few unexpected things come up in the last few weeks.

Overall, I think this is a really nice extension of MNE, and a useful tool for people to use for iEEG work. I have a few small comments / suggestions (something like a minor revision), but otherwise, it all looks good to me!

### General Comments

Following up on janfreyberg’s comment, I think the paper could still be clearer about what exactly this paper relates to, in terms of code. From my understanding, this contribution reflects the addition of a novel GUI for electrode localization and associated updated to make this work within the MNE. Somewhere around the first paragraph, I think this could just be explicitly stated ("this paper relates to the addition of a new ). Otherwise, without already knowing, reading the rest of the paper I find it a bit unclear what is new vs. existing.

I think it would also be useful to explicitly state which is the first version release of MNE to include this functionality (as in, what is the minimum version of MNE needed to use these utilities).

### Writing notes

Some small notes about the paper text:
- the first sentence of the first & second paragraph are basically the same. Could be edited for flow
- the figure legends repeat the figure number (“Figure 1: Figure 1.”)
- Figure 6A, the labels are a bit small and fuzzy, making it hard to read. Is it possible to tweak the size and/or resolution here?

### Code Review

In terms of code review, I primarily had a read through of the new [iEEG_locate_gui](https://github.com/mne-tools/mne-python/blob/main/mne/gui/_ieeg_locate_gui.py) file. Broadly, it all looks good to me, and also appears to have had a thorough review procedure in the associated [pull request](https://github.com/mne-tools/mne-python/pull/9586).

### Dependencies

In both the paper, and in the tutorials, I think it would be useful to explicitly list the dependencies that are required.

- Related, are all the dependencies for the GUI installed by default with MNE, or are any optional (requiring the user to install anything extra), and/or are there any particular settings the user would have to define to use the GUI? If so, these should be stated in the paper.

What are the core dependencies for this extension? `dipy` and `PyQT` are mentioned,  but other dependencies are used but not mentioned the paper (for example,  `nibabel`, which I think is used for managing / loading the anatomical files).

In practice, for example, to get the examples running locally, after installing the most recent release of MNE, I also had to separately install the following optional dependenices (of MNE) in order for this functionality to work:
- nibabel
- nilearn
- dipy
- mne_bids
- pyvistaqt**

This brings me to suggest to small updates:
- In the paper, I recommend explicitly noting all of these required dependencies, both so the reader has a sense of what will be needed in practice, and also, to acknowledge and cite the tools this extension is built on top of.
- In the tutorials, I recommend a note at the top listing any optional dependencies (beyond core MNE) that the tutorial requires, so it's explicit and the reader can install them all together before getting started.

In addition, `pooch` is required to install the example data, but has a sensible error message / suggested ToDo if it's not already installed, and I don't think necessarily needs anything adding about this.

**A note on pyvistaqt:
With a fresh install of MNE (for me at least) the backend defaulted to 'mayavi', and I initially got errors that mayavi is required. After installing this, I still get an uninformative error, until I called `mne.viz.set_3d_backend('pyvistaqt')` explicitly to set the backend. This is using the most recent stable version of MNE (0.24.1). I know `mayavi` is in the process of being deprecated from MNE, but it seems like with a fresh install of everything, it's plausible, maybe even default, to end up getting stuck with the wrong backend for the iEEG GUI, and it failing out uninformatively. At least while things are still a bit in flux, it might be worth adding an explicit note in the tutorial that `pyvistaqt` is the required backend, and note that this is therefore an extra dependency, and note the call needed to set this explicitly (`mne.viz.set_3d_backend('pyvistaqt')`).

Additionally, reading the tutorials notes that `freesurfer` is used for processing of the anatomical scans, and I think this could be explicitly noted and cited in the paper.

### Documentation

In terms of checking documentation and functionality, I read through and executed the three tutorial you listed as related to this paper.

Some brief notes on these tutorials:
- In tutorial `10_iEEG_localize` there are some notes on using freesurfer, for example, [here](https://github.com/mne-tools/mne-python/blob/main/tutorials/clinical/10_ieeg_localize.py#L104) which talks about running `recon-all`. First reading this, I thought I would have to run this long process to continue with the tutorial. It seems what is actually the case is that the downloaded example data is already processed (can be used to proceed with the tutorial), but that these notes are for anyone running their own data. This could be made explicit, I think ("When processing your own you will have to run XX. In this example we can proceed with the example data, which has already been processed.").
- At [this](https://github.com/mne-tools/mne-python/blob/main/tutorials/clinical/10_ieeg_localize.py#L279) line of the same tutorial, there is a syntax error (`.. note:` instead of `.. note::`), such that this note is not getting rendered on the website.
- Same tutorial, [here](https://github.com/mne-tools/mne-python/blob/main/tutorials/clinical/10_ieeg_localize.py#L420) there is a note about using "zooms" for speed here. I think this could use an extra sentence explaining what this means, as I find it unclear.

As a sidenote, I appreciate that you listed these tutorials in the paper, but only listing the names means that it's not necessarily obvious where to find these tutorials. Could you perhaps link each tutorial, and or link to the general tutorial page (something like "... are shown in the MNE-Python tutorials "X1", "X2", and "X3", available at https://mne.tools/stable/auto_tutorials/index.html").

### GUI

To my knowledge /  awareness, the code and GUI seem good and functional to me.

I have one small comment / note (though this might be a PyQT thing) - is there an easy way to allow for resizing individual panel sizes within the overall GUI? If I expand the size of the GUI overall, it seems the relative sizes of each sub-panel are effectively fixed, but in a way that the channel list is annoyingly large, and I would prefer to increase the size of the brain views - but it's not clear how. Is this adjustable?

Image of what I get when enlarging the GUI:
