### 2021-12-13 #3897 @alexrockhill
Thanks for the review @TomDonoghue, I think those comments were all very helpful and I will do a few pull requests in the coming days to add those changes.
