### 2021-12-19 #3897 @alexrockhill
> Otherwise, without already knowing, reading the rest of the paper I find it a bit unclear what is new vs. existing.

Just to clarify for the benefit of the reviewers with regards to this point (the text will be modified in the paper as well), all the functions that support the generation of the figures in the paper are new and the main contribution of the paper. Beyond that, for instance, `mne.viz.plot_3d_alignment` was modified heavily but not made from scratch, so it's a bit hard to completely delineate what is completely 100% new because the project was built on so many existing functions.
