### 2021-12-29 #3897 @leosunpsy
>

@alexrockhill @TomDonoghue The GUI resize problem has been fixed in my pull request ( #9633).
