### 2022-02-09 #3897 @alexrockhill
Does this PR mean that JOSS publications will be indexed on PubMed now? https://github.com/openjournals/joss-papers/pull/2831
