### 2022-02-11 #3897 @prashjha
Hi @alexrockhill, see if these suggestions make sense for your draft:

- Line 72: it should be "complete analysis and visualization of ..."?
- Lines 87-88, "Python-based libraries for image analysis and requires ..."?
- Lines 88-89, "MNE-Python code base benefits" does not make sense in this sentence.
- Line 102, "is implemented following coding best practices ..."?

After you have updated the draft, can you also do (if not done already) a 'tagged' release of your code
archive the release using zenedo or other methods? Make sure that the title of zenedo archive matches with the title of this JOSS submission.

Once you are done, I will hand your paper to EiC for the final decision.
