### 2022-02-12 #3897 @alexrockhill
We had added a few features since the images were generated (most notably lines for stereoelectrodes) so I updated the figures and did one last read-through for copy edits. I don't plan on any more changes, this is good to go by me.
