### 2022-02-12 #3897 @danielskatz
👋 @alexrockhill - I'm the AEiC on duty now, and have read your paper.  I've suggested a few small changes in https://github.com/mne-tools/mne-python/pull/10327 - please merge this or let me know what you disagree with, then we can continue to publishing.
