### 2022-02-12 #3897 @alexrockhill
Great thanks for the readover! I added everything except the Oxford commas. I appreciate your spending time to make corrections, thanks again.
