### 2022-02-14 #3897 @danielskatz
Assuming the only changes are in the acks, that looks fine.  Did you decide not to add the CZI part that was in https://github.com/mne-tools/mne-python/pull/10337?

Otherwise, we're just waiting for @janfreyberg to check off the last item, or for @prashjha to potentially agree that @janfreyberg meant to but didn't, if we don't hear back from @janfreyberg soon
