### 2022-02-14 #3897 @alexrockhill
No we didn't add CZI to the funding, the only person being funded by CZI is not on this paper, although it is a major funding source of MNE in general and so was mentioned on the paper. Thanks for checking!
