### 2022-02-14 #3897 @janfreyberg
> Does this PR mean that JOSS publications will be indexed on PubMed now? [openjournals/joss-papers#2831](https://github.com/openjournals/joss-papers/pull/2831)

I've asked on that issue, and the answer seems to be not yet. However, if JOSS is accepted in future, I believe all previously published papers also receive a PMID and are indexed. I would like to see this happen - it may be worth subscribing to the issue if this is of interest to you!
